% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{audit_formula}
\alias{audit_formula}
\title{Audit the Formula Used for Causal Inference}
\usage{
audit_formula(data, formula, exposure.name, extra.names)
}
\arguments{
\item{data}{Dataframe of raw data.}

\item{formula}{Formula with at least 1 exposure and 1 extra variable.}

\item{exposure.name}{The name of the exposure variable.}

\item{extra.names}{Character vector with the names of the extra variable.}
}
\value{
List with outcome.name, exposure.name and extra.names of variables.
}
\description{
Audit the formula used for causal inference.
}
\details{
The formula and the variables used must be validated to avoid problems.
All formula must have at least 1 outcome variable,
1 exposure variable and 1 extra variable. The extra variables can be confounds,
instrumental variable, etc. depending on the context.
}
\examples{
out <- audit_formula(
 iris, formula = Sepal.Length ~ Species * Sepal.Width,
 exposure.name = "Species", extra.names = "Sepal.Width")
# return a list with all 3 variable names
stopifnot(length(out) == 3)
}
