% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/backdr_dr_bad.R
\name{backdr_dr_bad}
\alias{backdr_dr_bad}
\alias{badstanddr}
\title{Doubly robust standardized estimates with misspecified outcome model}
\source{
section 6.3
}
\usage{
backdr_dr_bad(data, formula, exposure.name, confound.names)

badstanddr(data, formula, exposure.name, confound.names)
}
\arguments{
\item{data}{Dataframe of raw data.}

\item{formula}{Formula representing the model.}

\item{exposure.name}{Name of exposure variable.}

\item{confound.names}{Names of the confound variables.}
}
\value{
Dataframe in a useable format for \code{rsample::bootstraps}.
}
\description{
Doubly robust standardized estimates with misspecified outcome model.
}
\details{
Compute the doubly robust standardized estimates with erros in the outcome
model.
}
\examples{
# An example can be found in the location identified in the
# source section above at the github site
# https://github.com/FrankLef/FundamentalsCausalInference.
}
