% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/backdr_exp_np.R
\name{backdr_exp_bbX}
\alias{backdr_exp_bbX}
\title{Compute standardized averages using exposure modeling, non-parametric}
\usage{
backdr_exp_bbX(
  data,
  outcome.name = "Y",
  exposure.name = "T",
  confound.names = "H",
  weights = "n"
)
}
\arguments{
\item{data}{Dataframe of raw data.}

\item{outcome.name}{Name of outcome variable.}

\item{exposure.name}{Name of exposure variable.}

\item{confound.names}{Character vector of confound variable names.}

\item{weights}{String. Name of the columns with the weights that will
be used to create probabilities summing up to 1.}
}
\value{
List with 3 elements: EY1, EY0, EY0T1. See section 6.2.
for more details.
}
\description{
Compute standardized averages using exposure modeling, non-parametric.
}
\details{
Compute standardized averages using exposure modeling as described in
section 6.2.
}
