% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/backdr_exp_gee.R
\name{backdr_exp_gee}
\alias{backdr_exp_gee}
\alias{standexpgee}
\title{Compute standardized estimates with parametric exposure model}
\usage{
backdr_exp_gee(data, formula, exposure.name, confound.names)

standexpgee(data, formula, exposure.name, confound.names)
}
\arguments{
\item{data}{Dataframe of raw data.}

\item{formula}{Formula representing the model.}

\item{exposure.name}{Name of exposure variable.}

\item{confound.names}{Names of the confound variables.}
}
\value{
Dataframe in a useable format for \code{rsample::bootstraps}.
}
\description{
Compute standardized estimates with parametric exposure model.
}
\details{
The standardized estimates are computed using the exposure model and the
\code{geeglm} from the \code{geepack} package.
This method requires 2 different formulas which are created from the
arguments \code{formula}. The 2 formulas created are for the exposure model
and another one for the weighted linear model.
}
\seealso{
backdr_exp
}
