% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/backdr_exp_gee.R
\name{backdr_exp_geeX}
\alias{backdr_exp_geeX}
\title{Compute standardized estimates with parametric exposure model}
\usage{
backdr_exp_geeX(data, formula = Y ~ T + H, R = 1000, conf = 0.95)
}
\arguments{
\item{data}{Dataframe of raw data.}

\item{formula}{Formula must be in the form \code{Y ~ `T` + H}}

\item{R}{Number of bootstrap replicates.}

\item{conf}{Confidence interval.}
}
\value{
Dataframe of estimates
}
\description{
Compute standardized estimates with parametric exposure model.
}
\details{
The standardized estimates are computed using the exposure model and the
\code{geeglm} from the \code{gee} package.
This method requires 2 different formulas which are created from the
arguments \code{formula}. The 2 formulas created are for the exposure model
and another one for the weighted linear model.
}
\seealso{
backdr_exp
}
