% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/backdr_exp_np.R
\name{backdr_exp_np}
\alias{backdr_exp_np}
\alias{mk.mortdat}
\alias{attsem.r}
\title{Compute Standardized Averages Using Exposure Modeling, Non Parametric}
\source{
Section 6.2.1
}
\usage{
backdr_exp_np(data, formula, exposure.name, confound.names, att = FALSE)

mk.mortdat(data, formula, exposure.name, confound.names, att = FALSE)

attsem.r(data, formula, exposure.name, confound.names, att = FALSE)
}
\arguments{
\item{data}{Dataframe of raw data.}

\item{formula}{Formula representing the model.}

\item{exposure.name}{Name of exposure variable.}

\item{confound.names}{Names of the confound variables.}

\item{att}{if \code{FALSE} calculate the standardized (unconfounded)
causal effect. If \code{TRUE} calculate the average effect of treatment
on the treated.}
}
\value{
Dataframe in a useable format for \code{rsample::bootstraps}.
}
\description{
Compute standardized averages using exposure modeling, non-parametric.
}
\details{
Compute standardized averages using exposure modeling. See introduction of
section 6.2 and section 6.2.1.
}
