% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/backdr_exp.R
\name{backdr_expX}
\alias{backdr_expX}
\title{Compute standardized estimates with parametric exposure model}
\usage{
backdr_expX(data, formula = Y ~ T + H, att = FALSE, R = 1000, conf = 0.95)
}
\arguments{
\item{data}{Dataframe of raw data.}

\item{formula}{Formula must be in the form \code{Y ~ `T` + ...}}

\item{att}{if \code{FALSE} calculate the standardized (unconfounded)
causal effect. If \code{TRUE} calculate the average effect of treatment
on the treated.}

\item{R}{Number of bootstrap replicates.}

\item{conf}{Confidence interval.}
}
\value{
Dataframe of estimates using exposure-model standardization
}
\description{
Compute standardized estimates with parametric exposure model.
}
\details{
The standardized estimates are computed using the exposure model.
This method requires 2 different formulas which are created from the
arguments \code{formula}. The 2 formulas created are for the exposure model
and another one for the weighted linear model. Also \code{T}, i.e. the
exposure, must always be binary, if not it can be made binary
"one can first recode the data so that T = 1 when it is previously equaled,
and T = 0 when it previously equaled any value other than t", p. 113.
}
