% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/backdr_out.R
\name{backdr_out}
\alias{backdr_out}
\alias{standout}
\title{Compute standardized estimates with parametric outcome model}
\usage{
backdr_out(
  data,
  formula = Y ~ T + A + H,
  exposure.name = "T",
  family = c("binomial", "poisson", "gaussian")
)

standout(
  data,
  formula = Y ~ T + A + H,
  exposure.name = "T",
  family = c("binomial", "poisson", "gaussian")
)
}
\arguments{
\item{data}{Dataframe of raw data.}

\item{formula}{Formula representing the model.}

\item{exposure.name}{Name of exposure variable.}

\item{family}{Name of distribution. Must be in
\code{c("binomial", "poisson", "gaussian")}.}
}
\value{
Dataframe of estimates using outcome-model standardization
}
\description{
Compute standardized estimates with parametric outcome model.
}
\details{
The standardized estimates are computed using the outcome model. See chapter
6 for details.
}
\seealso{
effect_measures
}
