% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/backdr_out_np.R
\name{backdr_out_np}
\alias{backdr_out_np}
\title{Standardized estimates via Outcome Modeling, Non-Parametric}
\usage{
backdr_out_np(
  data,
  outcome.name = "Y",
  exposure.name = "T",
  confound.names = "H",
  att = FALSE
)
}
\arguments{
\item{data}{Dataframe of data.}

\item{outcome.name}{Name of outcome variable.}

\item{exposure.name}{Name of exposure variable.}

\item{confound.names}{Name of confound variable.}

\item{att}{if \code{FALSE} calculate the standardized (unconfounded)
causal effect. If \code{TRUE} calculate the average effect of treatment
on the treated.}
}
\value{
Estimate using outcome-model standardization.
}
\description{
Standardized estimates via outcome modeling, non-parametric.
}
\details{
The standardized estimates are computed using the non-parametric outcome
model. The calculations are performed without regression.
}
