% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/backdr_out_np.R
\name{backdr_out_npr}
\alias{backdr_out_npr}
\alias{stand}
\title{Standardized estimates via Outcome Modeling, Non-Parametric Regression}
\usage{
backdr_out_npr(
  data,
  outcome.name = "Y",
  exposure.name = "T",
  confound.names = "H",
  interactions = list(c("T", "A")),
  att = FALSE
)

stand(
  data,
  outcome.name = "Y",
  exposure.name = "T",
  confound.names = "H",
  interactions = list(c("T", "A")),
  att = FALSE
)
}
\arguments{
\item{data}{Dataframe of data.}

\item{outcome.name}{Name of outcome variable.}

\item{exposure.name}{Name of exposure variable.}

\item{confound.names}{Name of confound variable.}

\item{interactions}{List of character vectors with the interactions.}

\item{att}{if \code{FALSE} calculate the standardized (unconfounded)
causal effect. If \code{TRUE} calculate the average effect of treatment
on the treated.}
}
\value{
Estimate using outcome-model standardization.
}
\description{
Standardized estimates via outcome modeling, non-parametric regression.
}
\details{
The standardized estimates are computed using the non-parametric outcome
model. IMPORTANT: The formula must be in the format \code{Y ~ T + ...} where the
exposure is in the first position \code{T}.
}
