% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/backdr_out_np.R
\name{backdr_out_npX}
\alias{backdr_out_npX}
\title{Standardized estimates via Outcome Modeling, Non-Parametric}
\usage{
backdr_out_npX(
  data,
  outcome,
  exposure,
  confound,
  att = FALSE,
  R = 1000,
  conf = 0.95
)
}
\arguments{
\item{data}{Dataframe of raw data.}

\item{outcome}{Outcome variable}

\item{exposure}{Exposure variable}

\item{confound}{Confound variable}

\item{att}{if \code{FALSE} calculate the standardized (unconfounded)
causal effect. If \code{TRUE} calculate the average effect of treatment
on the treated.}

\item{R}{Number of bootstrap replicates.}

\item{conf}{Confidence interval.}
}
\value{
Estimate using outcome-model standardization.
}
\description{
Standardized estimates via outcome modeling, non-parametric.
}
\details{
The standardized estimates are computed using the non-parametric outcome
model. The calculations are performed without regression.
}
