% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/backdr_out.R
\name{backdr_outX}
\alias{backdr_outX}
\title{Compute standardized estimates with parametric outcome model}
\usage{
backdr_outX(
  data,
  formula = Y ~ T + H,
  family = c("binomial", "poisson", "gaussian"),
  R = 1000,
  conf = 0.95
)
}
\arguments{
\item{data}{Dataframe of raw data.}

\item{formula}{Formula in format \code{Y ~ T + ...} see details above.}

\item{family}{Character. Name of the distribution. Default is "binomial".}

\item{R}{Number of bootstrap replicates.}

\item{conf}{Confidence interval.}
}
\value{
Estimate using outcome-model standardization
}
\description{
Compute standardized estimates with parametric outcome model.
}
\details{
The standardized estimates are computed using the outcome model.
IMPORTANT: The formula must be in the format \code{Y ~ T + ...} where T
is the exposure.
}
