% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/backdr_twoparts.R
\name{backdr_twoparts}
\alias{backdr_twoparts}
\title{Compute standardized estimates with the 2-parts model}
\usage{
backdr_twoparts(data, formula = Y ~ T + H, cond = "Z", R = 1000, conf = 0.95)
}
\arguments{
\item{data}{Dataframe of raw data.}

\item{formula}{Formula in format \code{Y ~ T + ...} T must represent the
exposure variable.}

\item{cond}{String. Name of the binary variable used by the 2-parts method
to differentiate the 2 parts of the model.}

\item{R}{Number of bootstrap replicates.}

\item{conf}{Confidence interval.}
}
\value{
Estimate using 2-parts standardization
}
\description{
Compute standardized estimates with the 2-parts model.
}
\details{
This method is detailed in exercise 4 of chapter 6. It is useful when there
are many zeros.  Note that the results don't match exactly those obtained
with outcome-mode or exposure-model standardization.
}
