% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/backdr_twoparts.R
\name{backdr_twoparts_old}
\alias{backdr_twoparts_old}
\title{Compute standardized estimates with the 2-parts model}
\usage{
backdr_twoparts_old(
  data,
  outcome.name = "Y",
  exposure.name = "T",
  confound.names = "H",
  condition.name = "Z"
)
}
\arguments{
\item{data}{Dataframe of raw data.}

\item{outcome.name}{Name of outcome variable.}

\item{exposure.name}{Name of exposure variable.}

\item{confound.names}{Name of confound variable.}

\item{condition.name}{Character vector of confound variable names.}
}
\value{
Dataframe in a useable format for \code{rsample::bootstraps}.
}
\description{
Compute standardized estimates with the 2-parts model.
}
\details{
This method is detailed in exercise 4 of chapter 6. It is useful when there
are many zeros.  Note that the results don't match exactly those obtained
with outcome-mode or exposure-model standardization.
}
