% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boot_utils.R
\name{boot_est}
\alias{boot_est}
\title{Get estimate and CI using a function as input}
\usage{
boot_est(
  data,
  func,
  times = 1000,
  alpha = 0.05,
  seed = NULL,
  transf = c("identity", "exp", "expit"),
  terms = NULL,
  ...
)
}
\arguments{
\item{data}{Dataframe of raw data.}

\item{func}{Function applied to data by bootstrapping.}

\item{times}{Number of bootstrap replicates. Default is 1000.}

\item{alpha}{Alpha used by percentile to give interval in
\code{c(alpha, 1- alpha)}.}

\item{seed}{Seed for random number generator. If \code{NULL} the seed itself
is random.}

\item{transf}{Type of conversion. Must be one of
\code{c("identity", "exp", "expit")}, default is \code{identity}.}

\item{terms}{Character vector used to reorder the rows using the
\code{term} column. When \code{NULL}, no reordering takes place.}

\item{...}{Other named arguments for \code{func}.}
}
\value{
Dataframe of estimates with confidence interval.
}
\description{
Get estimate and CI using a function as input.
}
\details{
Boostrap with a function and extra arguments as input. The function is called
\code{statistic} to remain constant with \code{boot::boot}.
}
\seealso{
effect_transf
}
