% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boot_utils.R
\name{boot_run}
\alias{boot_run}
\title{Bootstrap and generate a dataframe of estimates with CI}
\usage{
boot_run(data, func, times = 1000, alpha = 0.05, ...)
}
\arguments{
\item{data}{Dataframe of raw data.}

\item{func}{Function applied to data by bootstrapping.}

\item{times}{Number of bootstrap replicates. Default is 1000.}

\item{alpha}{Alpha used by percentile to give interval in
\code{c(alpha, 1- alpha)}.}

\item{...}{Other named arguments for \code{statistic}.}
}
\value{
Dataframe of estimates with CI.
}
\description{
Bootstrap and generate a dataframe of estimates with CI.
}
\details{
Generate a dataframe of estimates with the columns
\code{c("est", "conf", "lci", "uci")}. The \code{boot::boot} function is used
for bootstraping and the function \code{boot::boot.ci} is used to compute the
confidence interval.
}
\seealso{
boot::boot boot::boot.ci
}
