% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boot_utils.R
\name{boot_run_td}
\alias{boot_run_td}
\title{Bootstrapping Confidence Intervals with Tidyverse}
\usage{
boot_run_td(data, func, times = 1000, alpha = 0.05, ...)
}
\arguments{
\item{data}{Dataframe of raw data.}

\item{func}{Function applied to data by bootstrapping.}

\item{times}{Number of bootstrap replicates. Default is 1000.}

\item{alpha}{Alpha used by percentile to give interval in
\code{c(alpha, 1- alpha)}.}

\item{...}{Additional arguments used by \code{func}.}
}
\value{
Dataframe with term, .lower, .estimate, .upper, .alpha, .method
}
\description{
Bootstrapping confidence intervals with tidyverse.
}
\details{
Bootstrapping confidence intervals using the \code{rsample} package.
}
