% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boot_utils.R
\name{bootR_run}
\alias{bootR_run}
\title{Bootstrapping Confidence Intervals with Base R}
\usage{
bootR_run(data, func, times = 1000, alpha = 0.05, seed = NULL, ...)
}
\arguments{
\item{data}{Dataframe of raw data.}

\item{func}{Function applied to data by bootstrapping.}

\item{times}{Number of bootstrap replicates. Default is 1000.}

\item{alpha}{Alpha used by percentile to give interval in
\code{c(alpha, 1- alpha)}.}

\item{seed}{Seed for random number generator. If \code{NULL} the seed itself
is random.}

\item{...}{Other named arguments for \code{statistic}.}
}
\value{
Dataframe with term, .lower, .estimate, .upper, .alpha, .method.
}
\description{
Bootstrapping confidence intervals with base R.
}
\details{
Generate a dataframe of estimates using he \code{boot::boot} function for
bootstrapping and \code{boot::boot.ci} to compute the confidence intervals.
}
\seealso{
boot::boot boot::boot.ci
}
