% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boot_utils.R
\name{calc_effect_measures}
\alias{calc_effect_measures}
\title{Calculate the effect measures}
\usage{
calc_effect_measures(val0, val1, log = TRUE)
}
\arguments{
\item{val0}{Numeric, base value}

\item{val1}{Numeric, treated value}

\item{log}{If TRUE (default) compute the effect measures on the log scale.
If FALSE compute the effect measures on the natural scale.
The risk difference measure (RD) is always on the natural scale.}
}
\value{
Named numeric vector.
}
\description{
Calculate the effect measures.
}
\details{
Calculate the effect measures and put them in a named numeric() vector.
}
\examples{
val0 <- 0.25
val1 <- 0.5
# when log is required, we use default value
out <- calc_effect_measures(val0, val1)
stopifnot(all(is.finite(out)))
# when natural scale is required, we use log = FALSE
out <- calc_effect_measures(val0, val1, log = FALSE)
stopifnot(all(is.finite(out)))
}
