% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_prob.R
\name{calc_prob}
\alias{calc_prob}
\title{Calculate Probabilities}
\usage{
calc_prob(data, var.name, prob.name = "prob")
}
\arguments{
\item{data}{Dataframe.}

\item{var.name}{Name of variable to compute probabilities.}

\item{prob.name}{Name of column holding the probabilities' values.}
}
\value{
Dataframe with grouped variable \code{var.name} and probabilities in
\code{prob.nam}
}
\description{
Calculate probabilities.
}
\details{
Calculate the probabilities of a variable given a dataset.
}
\examples{
calc_prob(airquality, var.name = "Temp", prob.name = "prob")
}
\seealso{
calc_prob_cond
}
