% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_prob.R
\name{calc_prob_cond}
\alias{calc_prob_cond}
\title{Calculate Probabilities Conditional on Other Variables}
\usage{
calc_prob_cond(data, condition.names, var.name, prob.name = "prob")
}
\arguments{
\item{data}{Dataframe.}

\item{condition.names}{Character vector of condition variables.}

\item{var.name}{Name of variable to compute probabilities.}

\item{prob.name}{Name of column holding the probabilities' values.}
}
\value{
Dataframe with grouped variable \code{var.name} and probabilities in
\code{prob.nam}
}
\description{
Calculate probabilities conditional on other variables.
}
\details{
Calculate the empirical distribution of a variable by different
conditions/groups.
}
\examples{
calc_prob_cond(airquality, condition.names = "Month",
 var.name = "Temp", prob.name = "prob")
}
\seealso{
calc_prob
}
