% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data1.R
\docType{data}
\name{cogdat}
\alias{cogdat}
\title{Children's Oncology Group from Robins 1989}
\format{
Dataframe with 1190 rows and 4 variables.
\describe{
\item{A1}{1 = experimental therapy, 0 = standard therapy}
\item{H2}{1 = response data is available}
\item{A2}{1 = patient received bone marrow transplant}
\item{Y}{1 = patient surviving at two years}
}
}
\usage{
cogdat
}
\description{
See section 1.2.5. Data from Children's Oncology Group.
}
\keyword{datasets}
