% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/did_loglinear.R
\name{did_loglinear}
\alias{did_loglinear}
\title{Compute the DiD Estimator with a Loglinear Model}
\usage{
did_loglinear(
  data,
  formula = Y ~ A,
  varsY = c("VL0", "VL1"),
  R = 1000,
  conf = 0.95
)
}
\arguments{
\item{data}{Dataframe of raw data.}

\item{formula}{Formula in format \code{Y ~ A} where \code{A}
is the exposure.}

\item{varsY}{Names of the outcome variables.}

\item{R}{Number of bootstrap replicates. Default is 1000.}

\item{conf}{Confidence interval. Default is 0.95.}
}
\value{
Estimates using the difference in differences.
}
\description{
Compute the DiD Estimator with a loglinear model.
}
\details{
the DiD Estimator using the scripts from section 7.2, p. 141-142.
}
