% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/did_utils.R
\name{did_longer}
\alias{did_longer}
\title{Convert a data.frame to long format}
\usage{
did_longer(
  data,
  outcomes = c("Y0", "Y1"),
  outcome = "Y",
  names_to = "var",
  timevar = "time"
)
}
\arguments{
\item{data}{Dataframe of data to convert to long format.}

\item{outcomes}{Character vector of base and final variables.}

\item{outcome}{Name that will be used for the single outcome variable.}

\item{names_to}{Name that will be used for the name variable.}

\item{timevar}{Name that will be used for the time variable.}
}
\value{
Dataframe in long format.
}
\description{
Convert a data.frame to long format.
}
\details{
Convert a data.frame to long format. A tie column will be added with 0 for
baseline and 1 for final.
}
\seealso{
did_linear did_loglinear did_logistic
}
