% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/effect_utils.R
\name{effect_exp}
\alias{effect_exp}
\title{Exponentiate the effect measures}
\usage{
effect_exp(data, vars = c(RR = "logRR", `RR*` = "logRR*", OR = "logOR"))
}
\arguments{
\item{data}{Dataframe.╪}

\item{vars}{Character() of measure names.}
}
\value{
Dataframe of converted effects measures.
}
\description{
Exponentiate the effect measures.
}
\details{
Exponentiate each selected measure in the dataframe and replace its name
accordingly.
}
