% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/effect_utils.R
\name{effect_expit}
\alias{effect_expit}
\title{Inverse logit the effect measures}
\usage{
effect_expit(data, vars = c(P = "logitP"))
}
\arguments{
\item{data}{Dataframe.}

\item{vars}{Character() of measure names.}
}
\value{
Dataframe of converted effects measures.
}
\description{
Inverse logit the effect measures.
}
\details{
Inverse logit each selected measure in the dataframe and replace its name
accordingly.
}
