% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/effect_utils.R
\name{effect_inv}
\alias{effect_inv}
\title{Apply an inverse function to effect measures}
\usage{
effect_inv(data, inv = c("exp", "expit", "none"), vars)
}
\arguments{
\item{data}{Numeric vector to transform.}

\item{inv}{Name of the inverse function. Must be in
\code{c("exp", "expit", "none")}. default is \code{exp}.}

\item{vars}{Character vector identifying the variables. Use the
\code{effect_vars} function to create the character vector.}
}
\value{
Inversed transform of input.
}
\description{
Apply an inverse function to effect measures.
}
\details{
Apply an inverse function to the effect measures. The names of the
transformed variables are also change accordingly.
}
\seealso{
effect_vars effect_exp effect_expit
}
