% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/effect_utils.R
\name{effect_measures}
\alias{effect_measures}
\title{Calculate the effect measures}
\usage{
effect_measures(val0, val1, log = TRUE)
}
\arguments{
\item{val0}{Numeric, base value}

\item{val1}{Numeric, treated value}

\item{log}{If TRUE (default) compute the effect measures on the log scale.
If FALSE compute the effect measures on the natural scale.
The risk difference measure (RD) is always on the natural scale.}
}
\value{
Named numeric vector.
}
\description{
Calculate the effect measures.
}
\details{
Calculate the effect measures and put them in a named numeric() vector.
If \code{log = TRUE} the measures are RD (risk difference),
logRR (log of relative risk), logRR* (log of other relative risk) and
logOR (log of odds ratio). If \code{log = FALSE} the measures are
RD (risk difference), logRR (relative risk),
logRR* (other relative risk) and logOR (odds ratio). See section 3.3 for
more details on these measures.
}
\examples{
val0 <- 0.25
val1 <- 0.5
# when log is required, we use default value
out <- effect_measures(val0, val1)
stopifnot(all(is.finite(out)))
# when natural scale is required, we use log = FALSE
out <- effect_measures(val0, val1, log = FALSE)
stopifnot(all(is.finite(out)))
}
