% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/effect_transf.R
\name{effect_transf}
\alias{effect_transf}
\title{Convert Dataframe of Effect Measures to Its Inverse}
\usage{
effect_transf(data, type = c("identity", "exp", "expit"))
}
\arguments{
\item{data}{Dataframe of effect measures.}

\item{type}{Type of conversion. Must be one of
\code{c("identity", "exp", "expit")}, default is \code{identity}}
}
\value{
Data.frame
}
\description{
#' Convert dataframe of effect measures to its inverse.
}
\details{
The \code{.estimate}, \code{.lower} and \code{.upper} will be trabsformed
and the \code{term} will be renamed.
}
