% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/effect_transf.R
\name{effect_transf_proc}
\alias{effect_transf_proc}
\title{Convert Dataframe of Effect Measures to Its Inverse}
\usage{
effect_transf_proc(data, prefix = c("log", "logit"))
}
\arguments{
\item{data}{Dataframe of effect measures.}

\item{prefix}{Type of conversion. Must be one of
\code{c("log", "logit")}.}
}
\value{
Data.frame
}
\description{
#' Convert dataframe of effect measures to its inverse.
}
\details{
The \code{.estimate}, \code{.lower} and \code{.upper} will be trabsformed
and the \code{term} will be renamed.
}
