% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/effect_utils.R
\name{effect_vars}
\alias{effect_vars}
\title{Charatcer vector of measure names and their inverse}
\usage{
effect_vars(evars = c("standard", "modifier", "logit"))
}
\arguments{
\item{evars}{String. Should one of \code{c("standard", "modifier", "logit")}.
Default value is \code{"standard"}. It is the set of terms to use with the
inverse function defined by \code{inv}. For example}
}
\value{
Named character vector.
}
\description{
Charatcer vector of measure names and their inverse.
}
\details{
Create charatcer vector of measure names and their inverse to be used with
the output of effect measures and the inverse functions.
}
\seealso{
effect_measures
}
