% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frontdr.R
\name{est_frontdr_np}
\alias{est_frontdr_np}
\title{Estimate the effect using Front-Door Method}
\usage{
est_frontdr_np(data, outcome, exposure, surrogate)
}
\arguments{
\item{data}{Dataframe of raw data.}

\item{outcome}{Name of outcome variable.}

\item{exposure}{Name of exposure variable.}

\item{surrogate}{Name of surrogate variable.}
}
\value{
Numeric vector.
}
\description{
Estimate the effect using Front-door method.
}
\details{
Estimate the effect using Front-door method. See chap. 8, section 8.3
for details.
}
