% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boot_utils.R
\name{exp_effects}
\alias{exp_effects}
\title{Exponentiate the effect measures}
\usage{
exp_effects(data, vars = c(RR = "logRR", `RR*` = "logRR*", OR = "logOR"))
}
\arguments{
\item{data}{Dataframe.}

\item{vars}{Character() of measure names.}
}
\value{
Dataframe of converted effects measures.
}
\description{
Exponentiate the effect measures.
}
\details{
Exponentiate each selected measure in the dataframe and replace its name
accordingly.
}
