% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data2.R
\docType{data}
\name{fci_sim_08_01}
\alias{fci_sim_08_01}
\title{Simulation 8.1}
\format{
Dataframe with 13 rows and 7 variables.
\describe{
\item{A}{Exposure}
\item{S}{Surrogate marker}
\item{Y}{Outcome}
\item{Ydot0}{Potential outcome Y(., S(0))}
\item{Ydot1}{Potential outcome Y(., S(1))}
}
}
\usage{
fci_sim_08_01
}
\description{
See chapter 8, section 8.3.
}
\keyword{datasets}
