% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data2.R
\docType{data}
\name{fci_tbl_04_02}
\alias{fci_tbl_04_02}
\title{Table 4.2}
\format{
Dataframe with 18 rows and 6 variables.
\describe{
\item{term}{name of estimator}
\item{group}{grouping variable}
\item{.estimate}{estimate}
\item{.lower}{lower confidence interval}
\item{.upper}{upper confidence interval}
\item{.alpha}{alpha value confidence interval width = 1 - alpha}
\item{.method}{method used used to estimate interval}
}
}
\usage{
fci_tbl_04_02
}
\description{
See chapter 4.
}
\keyword{datasets}
