% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data2.R
\docType{data}
\name{fci_tbl_05_01}
\alias{fci_tbl_05_01}
\title{Table 5.1}
\format{
Dataframe with 1000 rows and 4 variables.
\describe{
\item{A}{exposure}
\item{Y0}{potential outcome 0}
\item{Y1}{potential outcome 1}
\item{Y}{outcome}
}
}
\usage{
fci_tbl_05_01
}
\description{
See section 5.2..
}
\keyword{datasets}
