% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data2.R
\docType{data}
\name{fci_tbl_06_01}
\alias{fci_tbl_06_01}
\title{Table 6.1}
\format{
Dataframe with 4 rows and 6 variables.
\describe{
\item{term}{name of estimator}
\item{.lower}{lower confidence interval}
\item{.estimate}{estimate}
\item{.upper}{upper confidence interval}
\item{.alpha}{alpha value confidence interval width = 1 - alpha}
\item{.method}{method used used to estimate interval}
}
}
\usage{
fci_tbl_06_01
}
\description{
See chapter 6.
}
\keyword{datasets}
