% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data2.R
\docType{data}
\name{fci_tbl_06_04}
\alias{fci_tbl_06_04}
\title{Table 6.4}
\format{
Dataframe with 4 rows and 5 variables.
\describe{
\item{name}{name of estimator}
\item{est}{estimate}
\item{conf}{confidence interval width}
\item{lci}{lower confidence interval}
\item{uci}{upper confidence interval}
}
}
\usage{
fci_tbl_06_04
}
\description{
See chapter 6.
}
\keyword{datasets}
