% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data2.R
\docType{data}
\name{fci_tbl_06_13}
\alias{fci_tbl_06_13}
\title{Table 6.3 and 6.14}
\format{
Dataframe with 4 rows and 6 variables.
\describe{
\item{ss}{nb ofcovariates}
\item{estimator}{estimate}
\item{description}{description of estimator}
\item{mean}{mean}
\item{sd}{standard deviation}
\item{pva}{p-value}
}
}
\usage{
fci_tbl_06_13
}
\description{
See chapter 6, section 6.3.
}
\keyword{datasets}
