% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data2.R
\docType{data}
\name{fci_tbl_07_02}
\alias{fci_tbl_07_02}
\title{Table 7.2 data}
\format{
Dataframe with 13 rows and 7 variables.
\describe{
\item{method}{method of estimation}
\item{name}{name of estimator}
\item{Truth}{true value}
\item{est}{estimate}
\item{conf}{confidence interval width}
\item{lci}{lower confidence interval}
\item{uci}{upper confidence interval}
}
}
\usage{
fci_tbl_07_02
}
\description{
See chapter 7, section 7.2.
}
\keyword{datasets}
