% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data2.R
\docType{data}
\name{fci_tbl_09_01a}
\alias{fci_tbl_09_01a}
\title{Table 9.1 using qt() instead of 1.96 for CI}
\format{
Dataframe with 10 rows and 7 variables.
\describe{
\item{method}{method of estimation}
\item{name}{name of estimator}
\item{Truth}{true value}
\item{est}{estimate}
\item{conf}{confidence interval width}
\item{lci}{lower confidence interval}
\item{uci}{upper confidence interval}
}
}
\usage{
fci_tbl_09_01a
}
\description{
See chapter 9, section 9.3.
}
\keyword{datasets}
