% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boot_utils.R
\name{formula2vars}
\alias{formula2vars}
\title{Extract the name of variables and coefficients from a formula}
\usage{
formula2vars(formula)
}
\arguments{
\item{formula}{Object of class \code{formula}.}
}
\value{
List of variable names
}
\description{
Extract the name of variables and coefficients from a formula
}
\examples{
formula2vars(Y ~ A + X1*X2)
}
