% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frm_parse.R
\name{frm_parse}
\alias{frm_parse}
\title{Parse and validate formula for estimate.}
\usage{
frm_parse(frm)
}
\arguments{
\item{frm}{String or formula to parse}
}
\value{
Object of class \code{Formula}
}
\description{
Parse and validate formula for estimate.
}
