% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frm_parse.R
\name{frm_parse_check}
\alias{frm_parse_check}
\title{Validate formula as a whole.}
\usage{
frm_parse_check(frm)
}
\arguments{
\item{frm}{String or formula to parse}
}
\value{
Object of class \code{Formula}
}
\description{
Validate formula as a whole.
}
