% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frm_parse.R
\name{frm_parse_rgx}
\alias{frm_parse_rgx}
\title{Validate formula's text patterns}
\usage{
frm_parse_rgx(frm)
}
\arguments{
\item{frm}{String or formula to parse}
}
\value{
Object of class \code{Formula}
}
\description{
Validate formula's text patterns
}
