% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frontdr.R
\name{frontdr_np}
\alias{frontdr_np}
\alias{frontdoor.r}
\title{Estimate the Effect Using the Front-Door Method}
\source{
Section 8.3
}
\usage{
frontdr_np(data, formula, exposure.name, surrogate.name)

frontdoor.r(data, formula, exposure.name, surrogate.name)
}
\arguments{
\item{data}{Dataframe of raw data.}

\item{formula}{Formula representing the model.}

\item{exposure.name}{Name of exposure variable.}

\item{surrogate.name}{Name of the surrogate variable.}
}
\value{
Dataframe in a useable format for \code{rsample::bootstraps}.
}
\description{
Estimate the effect using Front-door method.
}
\details{
Estimate the effect using Front-door method. The exposure variable is used
as a confounder in the front-door method.
}
\examples{
# An example can be found in the location identified in the
# source section above at the github site
# https://github.com/FrankLef/FundamentalsCausalInference.
}
