% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggp_dag.R
\name{ggp_dag}
\alias{ggp_dag}
\title{Plot a DAG with ggplot and node names with subscripts}
\source{
https://www.r-bloggers.com/2019/10/how-to-use-math-symbols-with-ggdag-2/
}
\usage{
ggp_dag(
  dag,
  layout = "sugiyama",
  text_labels = NULL,
  text_size = 7,
  box_padding = 1,
  legend_pos = "bottom"
)
}
\arguments{
\item{dag}{dag object}

\item{layout}{Name of layout, default is "sugiyama".}

\item{text_labels}{If NULL (default) the nodes' names are used as is. If
a vector of text labels is given they MUST BE IN ALPHABETICAL ORDER.}

\item{text_size}{Integer. Size of the nodes' name.}

\item{box_padding}{Box padding for \code{geom_dag_label_repel}}

\item{legend_pos}{Legend position of ggplot.}
}
\value{
ggplot object
}
\description{
Plot a DAG with ggplot and node names with subscripts.
}
\details{
The DAG will be plotted using \code{ggplot2} and the nodes' names can have
subscript as in mathematical notations. See the \code{source} below for
more details.
}
