% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggp_format.R
\name{ggp_format}
\alias{ggp_format}
\title{Create a list of argumaents used by ggp_x custom functions}
\usage{
ggp_format(data, type = c("none", "measures_tbl"))
}
\arguments{
\item{data}{Data object.}

\item{type}{String to identify the desired set of arguments.}
}
\value{
List of argments including \code{data}
}
\description{
Create a list of argumaents used by ggp_x custom functions
}
