% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggp_measures.R
\name{ggp_measures}
\alias{ggp_measures}
\title{Plot of effect measures}
\usage{
ggp_measures(
  data,
  title = "Title",
  subtitle = "Subtitle",
  vline = list(colors = c("lightseagreen", "violet"), linetype = "solid", size = 3, alpha
    = 0.5),
  pointrange = list(size = 1, fatten = 2),
  text = list(size = 3, color = "navy", digits = 2),
  scale_color = list(zero = "darkgreen", one = "magenta"),
  text_size = list(title = 0.9, y_axis = 0.9)
)
}
\arguments{
\item{data}{Data frame with effect measures.}

\item{title}{Title.}

\item{subtitle}{Subtitle.}

\item{vline}{geom_vline specs.}

\item{pointrange}{geom_pointrange specs.}

\item{text}{geom_text_repel specs.}

\item{scale_color}{scale_color_manual color specs.}

\item{text_size}{text size for title and axis as a factor of base font size.}
}
\value{
A ggplot object.
}
\description{
Plot of effect measures.
}
\details{
Plot of effect measures with causal power (CF) and attributable fraction
(AF). See secion 3.3, p. 44-45 for more details.
}
