% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggp_measures.R
\name{ggp_measures_df}
\alias{ggp_measures_df}
\title{Create the data fame for \code{ggp_measures}}
\usage{
ggp_measures_df(data)
}
\arguments{
\item{data}{Data frame to plot.}
}
\value{
Data frame.
}
\description{
Create the data fame for \code{ggp_measures}
}
\seealso{
ggp_measures
}
