% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggp_measures_modif.R
\name{ggp_measures_modif}
\alias{ggp_measures_modif}
\title{Create a plot of effect-measures modifications}
\usage{
ggp_measures_modif(
  df,
  title = "Change in effect measures",
  subtitle = "Effect-Measure Modification and Statistical Interaction"
)
}
\arguments{
\item{df}{Dataframe to plot.}

\item{title}{Title of the plot.}

\item{subtitle}{Subtitle of the plot.}
}
\value{
A ggplot object.
}
\description{
Create a plot of effect-measures modifications
}
\details{
Create a plot of effect-measures modifications. The measures are shoen
as solid lines. The actual effects are shown in dashed lines.
}
