% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data1.R
\docType{data}
\name{gss}
\alias{gss}
\title{Dataset from 2018 GSS}
\format{
Dataframe with 2348 rows and 12 variables.
\describe{
\item{age}{age of participant.}
\item{gt65}{1 = age greater than 65.}
\item{attend}{1 = attending religious service more than once a month.}
\item{gthsedu}{1 = greater than a high school education.}
\item{magthsedu}{1 = participant's mother having greater than a
high school education.}
\item{pagthsedu}{1 = participant's father having greater than a
high school education.}
\item{fair}{1 = belief that if people got a chance they would try to be fair.}
\item{owngun}{1 = owning a gun or revolver.}
\item{conservative}{1 = political views that are slighlty conservative or more.}
\item{trump}{0 = those who did not vote or did not answer are coded 0.}
\item{white}{1 = participant considers himself white.}
\item{female}{respondent's sex}
}
an exported object
}
\source{
\url{www.gss.norc.org}
}
\usage{
gss
}
\description{
See section 1.2.4. Data on 2348 surveyed individuals.
}
\keyword{datasets}
