% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gt_utils.R
\name{gt2ggp}
\alias{gt2ggp}
\title{Convert \code{gt_tbl} to \code{ggplot}}
\usage{
gt2ggp(obj, ext = ".png", ...)
}
\arguments{
\item{obj}{Object of class \code{gt_tbl}.}

\item{ext}{Image file exension. Default is ".png".}

\item{...}{Additional arguments for \code{gt::gtsave}.}
}
\value{
Object of class \code{ggplot}.
}
\description{
Convert \code{gt_tbl} to \code{ggplot}
}
