% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gt_utils.R
\name{gt_basic}
\alias{gt_basic}
\title{Basic format of table created with \code{gt}}
\usage{
gt_basic(tbl, title, subtitle)
}
\arguments{
\item{tbl}{Object of class \code{gt_tbl}.}

\item{title}{String. Table title.}

\item{subtitle}{String. Table subcols_align title.}
}
\value{
Formatted \code{gt_tbl} object.
}
\description{
Basic format of table created with \code{gt}.
}
\details{
Format of table created with \code{gt} for the heading, row stripping,
source note, etc.
}
