% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gt_measures.R
\name{gt_measures}
\alias{gt_measures}
\title{Create a table of effect measures with their CI}
\usage{
gt_measures(
  df,
  conf = df$conf[1],
  digits = 3,
  title = "Title",
  subtitle = "Subtitle"
)
}
\arguments{
\item{df}{Dataframe of measures.}

\item{conf}{Numeric, confidence interval.}

\item{digits}{Integer, number of digits to the right of the decimal.}

\item{title}{The title of the table.}

\item{subtitle}{Subtitle of the table.}
}
\value{
A gt object create by the \code{gt} package
}
\description{
Create a table of effect measures with their CI.
}
\details{
The table will have 3 columns. The measure name, the estimates and the
confidence interval.
}
