% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gt_measures_colgrp.R
\name{gt_measures_colgrp}
\alias{gt_measures_colgrp}
\title{Create a table of effect-measure modifications with their CI}
\usage{
gt_measures_colgrp(
  df,
  var_grp = "group",
  digits = 3,
  title = "Title",
  subtitle = "Subtitle"
)
}
\arguments{
\item{df}{Dataframe of results.}

\item{var_grp}{Name of variable used to group the columns.}

\item{digits}{Integer, number of digits to the right of the decimal.}

\item{title}{The title of the table.}

\item{subtitle}{Subtitle of the table.}
}
\value{
A \code{gt_tbl} object
}
\description{
Create a table of effect-measure modifications with their CI.
}
\details{
The table will have 3 columns for each of the 2 strata and the
difference column. The measure name, the estimates and the
confidence interval.
}
