% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gt_probs.R
\name{gt_probs}
\alias{gt_probs}
\title{Create table of probabilities with \code{gt} package}
\usage{
gt_probs(df, digits = 3, title, subtitle)
}
\arguments{
\item{df}{Datafframe}

\item{digits}{Integer, number of digits to the right of decimal}

\item{title}{Title of the table}

\item{subtitle}{Subtitle of the table}
}
\value{
A \code{gt_tbl} object
}
\description{
Create table of probabilities with \code{gt} package.
}
\details{
Create table of probabilities with \code{gt} package. The last column
represents the probabilities.
}
