% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gt_standdr.R
\name{gt_standdr}
\alias{gt_standdr}
\title{Create a table from the result of simulating doubly robust}
\usage{
gt_standdr(df, title = "Title", subtitle = "Subtitle")
}
\arguments{
\item{df}{Dataframe of measures.}

\item{title}{The title of the table.}

\item{subtitle}{Subtitle of the table.}
}
\value{
A \code{gt_tbl} object
}
\description{
Create a table from the result of simulating doubly robust.
}
\details{
The table shows the result by grouping the ss with separate column groups.
}
