% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/instr_logistic.R
\name{instr_logistic}
\alias{instr_logistic}
\alias{ivlogit.r}
\title{Estimate Effect Using Instrument Variables via Logistic Fit}
\source{
Section 9.3
}
\usage{
instr_logistic(
  data,
  formula,
  exposure.name,
  instrument.name,
  tol = .Machine$double.eps^0.5,
  niter = 10L
)

ivlogit.r(
  data,
  formula,
  exposure.name,
  instrument.name,
  tol = .Machine$double.eps^0.5,
  niter = 10L
)
}
\arguments{
\item{data}{Dataframe of raw data.}

\item{formula}{Formula representing the model. It must be a saturated model,
e.g. \code{Y ~ T * A}.}

\item{exposure.name}{Name of exposure variable.}

\item{instrument.name}{Name of instrument variable.}

\item{tol}{Numeric > 0. Tolerance used in estimation. Default is
.Machine$double.eps^0.5.}

\item{niter}{Number of iterations}
}
\value{
Dataframe in a useable format for \code{rsample::bootstraps}.
}
\description{
Estimate effect using instrument variables via logistic fit.
}
\details{
See section 9.3 p. 167-168 for details. \bold{It is important to note that the
formula must specify a saturated model, e.g. Y ~ T * A}.
}
\examples{
# An example can be found in the location identified in the
# source section above at the github site
# https://github.com/FrankLef/FundamentalsCausalInference.
}
