% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/instr_loglinear.R
\name{instr_loglinear}
\alias{instr_loglinear}
\title{Estimate Effect Using Instrument Variables via Logarithmic fit}
\usage{
instr_loglinear(
  data,
  outcome.name = "Y",
  exposure.name = "A",
  instrument.name = "T",
  tol = .Machine$double.eps^0.5,
  niter = 10L
)
}
\arguments{
\item{data}{Dataframe of raw data.}

\item{outcome.name}{Name of outcome variable.}

\item{exposure.name}{Name of exposure variable.}

\item{instrument.name}{Name of instrument variable.}

\item{tol}{Numeric > 0. Tolerance used in estimation.}

\item{niter}{Number of iterations}
}
\value{
Numeric vector of estimates.
}
\description{
Estimate effect using instrument variables via logarithmic fit.
}
\details{
See section 9.3 p. 167-168 for details.
}
