% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boot_utils.R
\name{invlogit_effects}
\alias{invlogit_effects}
\title{Inverse logit the effect measures}
\usage{
invlogit_effects(data, vars = c(P = "logitP"))
}
\arguments{
\item{data}{Dataframe.}

\item{vars}{Character() of measure names.}
}
\value{
Dataframe of converted effects measures.
}
\description{
Inverse logit the effect measures.
}
\details{
Inverse logit each selected measure in the dataframe and replace its name
accordingly.
}
