% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jack_utils.R
\name{jack_ci}
\alias{jack_ci}
\title{Compute the Confidence Interval Estimated with Jacknife}
\usage{
jack_ci(x, conf = 0.95)
}
\arguments{
\item{x}{Numerical vector of effect measure}

\item{conf}{Confidence interval width. Default is 0.95.}
}
\value{
Numeric vector with estimates and confidence interval.
}
\description{
Compute the confidence interval estimated with jacknife.
}
\details{
This function is used to calculate the confidence interval using input
data estimated with the jacknife method.
}
