% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jack_utils.R
\name{jack_est}
\alias{jack_est}
\title{Estimate of Effect Measure and CI With Jacknife (LOO)}
\usage{
jack_est(
  data,
  func,
  conf = 0.95,
  inv = c("exp", "expit", "none"),
  evars = c("standard", "modifier", "logit"),
  ...
)
}
\arguments{
\item{data}{Dataframe of raw data.}

\item{func}{Function to estimate the effect measure.}

\item{conf}{Confidence interval width. Default is 0.95.}

\item{inv}{Choice of inverse function to apply to the effect measure
\code{exp} will exponentiate the result (Default), \code{expit} will apply
the inverse logit and \code{none} will do nothing (identity function).}

\item{evars}{String. Should one of \code{c("standard", "modifier", "logit")}.
Default value is \code{"standard"}.}

\item{...}{Other named arguments for \code{func}.}
}
\value{
Dataframe of estimates with confidence interval.
}
\description{
Estimate of effect measure and CI With jacknife (LOO).
}
\details{
Compute the confidence interval of effect measures using the jacknife method.
It uses the \code{loo_cv} function from the \code{rsample} package. The
\code{jacknife} funciton from the \code{resample} is not maintained anymore
and the author of \code{resample} recommend using the \code{rsample} package.
The output is transformed with inverse function when required. See section
9.3 of the book for more details.
}
\seealso{
jack_run jack_ci
}
