% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mc_standdr.R
\name{mc_standdr}
\alias{mc_standdr}
\title{Monte Carlo Simulation of Doubly Robust Standardization}
\source{
Section 6.3.1
}
\usage{
mc_standdr(ss = c(40, 100), nrep = 1000, width = 0.95)
}
\arguments{
\item{ss}{Integer(). Number of covariates.}

\item{nrep}{Number of Monte Carlo repetitions.}

\item{width}{Width of interval. e.g. 0.95 will give interval c(0.025, 0.975).
Default is 0.95.}
}
\value{
Dataframe of results.
}
\description{
Monte Carlo Simulation of Doubly Robust Standardization
}
\seealso{
standdr_sim standdr_est
}
