% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meas_effect_cond.R
\name{meas_effect_cond}
\alias{meas_effect_cond}
\alias{bootc}
\title{Compute estimates of the conditional association measures}
\usage{
meas_effect_cond(
  data,
  formula,
  exposure.name,
  confound.names,
  condition.names = confound.names,
  family = c("binomial", "poisson", "gaussian")
)

bootc(
  data,
  formula,
  exposure.name,
  confound.names,
  condition.names = confound.names,
  family = c("binomial", "poisson", "gaussian")
)
}
\arguments{
\item{data}{Dataframe of raw data.}

\item{formula}{The model formula.}

\item{exposure.name}{Name of the exposure variable.}

\item{confound.names}{Name of confound variables.}

\item{condition.names}{Character vector of conditioned variable names. By
default, it will be the same as the \code{confound.names}.}

\item{family}{Name of distribution. Must be in
\code{c("binomial", "poisson", "gaussian")}.}
}
\value{
Dataframe in a useable format for \code{rsample::bootstraps}.
}
\description{
Compute estimates of the conditional association measures.
}
\details{
Estimate the expected conditional outcomes and the
conditional effect or association measures. See exercise 3 of chapter 3
for a full example on how to use this function.
}
