% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meas_effect_modif.R
\name{meas_effect_modif}
\alias{meas_effect_modif}
\alias{boot}
\title{Compute estimates of the association measures for 2 strata}
\usage{
meas_effect_modif(data, formula = Y ~ T + M, R = 1000, conf = 0.95)

boot(data, formula = Y ~ T + M, R = 1000, conf = 0.95)
}
\arguments{
\item{data}{Dataframe of raw data.}

\item{formula}{Formula in format \code{Y ~ `T` + M}}

\item{R}{Number of bootstrap replicates.}

\item{conf}{Confidence interval}
}
\value{
Dataframe of summarized results
}
\description{
Compute estimates of the association measures for 2 strata.
}
\details{
IMPORTANT:
The formula must always be in form Y ~ \code{T} + M, that is with only
2 predictors: T as the treatment variable and M as the modifier
variable
}
