% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mediation.R
\name{mediation}
\alias{mediation}
\title{Estimate Mediation Effect with Parametric Assumptions}
\source{
section 12.3
}
\usage{
mediation(data, formula, exposure.name, mediator.name, confound.names)
}
\arguments{
\item{data}{Dataframe of raw data.}

\item{formula}{Formula representing the model.}

\item{exposure.name}{Name of exposure variable.}

\item{mediator.name}{Name of mediator variable.}

\item{confound.names}{Names of the confound variables.}
}
\value{
Dataframe in a useable format for \code{rsample::bootstraps}.
}
\description{
Estimate mediation effect with parametric assumptions.
}
\details{
Estimate mediation effect with parametric assumptions using the
\code{tidyverse} way of coding.
}
\examples{
# An example can be found in the location identified in the
# source section above at the github site
# https://github.com/FrankLef/FundamentalsCausalInference.
}
\seealso{
mediation_calc
}
