% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mediation_calc.R
\name{mediation_calc}
\alias{mediation_calc}
\title{Calculate the mediation variables.}
\usage{
mediation_calc(NDE, NIE, CDE0, CDE1)
}
\arguments{
\item{NDE}{Number. Natural direct effect.}

\item{NIE}{Number. Natural indirect effect.}

\item{CDE0}{Number. Controlled direct effect setting M=0.}

\item{CDE1}{Number. Controlled direct effect setting M=1.}
}
\value{
Dataframe in a useable format for \code{rsample::bootstraps}.
}
\description{
Calculate the mediation variables.
}
\details{
Calculate the mediation variables and put them in a dataframe that can be
used with bootstrapping.
}
\examples{
mediation_calc(NDE = -0.03552486, NIE = -0.11595150,
  CDE0 = 0.04240090, CDE1 = 0.04705889)
}
\seealso{
mediation
}
