% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mediation_NIE.R
\name{mediation_NIE}
\alias{mediation_NIE}
\title{Estimate the Natural Indirect Effect of a Mediator Variable}
\source{
section 12.3
}
\usage{
mediation_NIE(data, formula, exposure.name, mediator.name, confound.names)
}
\arguments{
\item{data}{Dataframe of raw data.}

\item{formula}{Formula representing the model.}

\item{exposure.name}{Name of exposure variable.}

\item{mediator.name}{Name of mediator variable.}

\item{confound.names}{Names of the confound variables.}
}
\value{
Dataframe in a useable format for \code{rsample::bootstraps}.
}
\description{
Estimate the natural indirect effect of a mediator variable.
}
\details{
Estimate the natural indirect effect of a mediator variable with parametric
assumptions.
}
\examples{
# An example can be found in the location identified in the
# source section above at the github site
# https://github.com/FrankLef/FundamentalsCausalInference.
}
