% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mortality}
\alias{mortality}
\title{Mortality Rates by Age and Country}
\format{
Dataframe with 4 rows and 5 variables.
\describe{
\item{T}{1 = US, 0 = China}
\item{H}{1 = > 65 years old}
\item{deaths}{nb of deaths}
\item{population}{size of population}
\item{Y}{Mortality Rate = deaths / population}
}
}
\usage{
mortality
}
\description{
See section 1.2.1. Mortality rates by age group in the US and China in
2019.
}
\keyword{datasets}
