% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data1.R
\docType{data}
\name{mortality_long}
\alias{mortality_long}
\title{Mortality Rates by Age and Country in long format.}
\format{
Dataframe with 4 rows and 5 variables.
\describe{
\item{T}{1 = US, 0 = China}
\item{H}{1 = > 65 years old}
\item{Y}{ 1 = deaths proportion, 0 = living proportion}
\item{population}{size of population}
\item{p}{proportion}
}
}
\usage{
mortality_long
}
\description{
See section 1.2.1. Mortality rates by age group in the US and China in
2019.
}
\keyword{datasets}
