% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data1.R
\docType{data}
\name{nces}
\alias{nces}
\title{Admissions data from the NCES IPEDS 2018-2019 provisionally.}
\format{
Dataframe with 1217 rows and 3 variables.
\describe{
\item{selective}{1 = institution admitted less than 50\% of applicants}
\item{female}{1 = more than 69\% of students admitted were women}
\item{highmathsat}{1 = average of the 25th and 75th precentiles of SAT was
higher than 600}
}
}
\source{
\url{nces.ed.gov}
}
\usage{
nces
}
\description{
See section 1.2.2 for details. Admissions data collected in the Fall 2018.
}
\keyword{datasets}
