% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/precision.R
\name{precision_eff}
\alias{precision_eff}
\alias{precision.r}
\title{Compute Precision efficiency}
\source{
Section 11.2
}
\usage{
precision_eff(data, formula, exposure.name, precision.name, none.name = "None")

precision.r(data, formula, exposure.name, precision.name, none.name = "None")
}
\arguments{
\item{data}{Dataframe of raw data.}

\item{formula}{Formula representing the model.}

\item{exposure.name}{Name of exposure variable.}

\item{precision.name}{Name of precision variable.}

\item{none.name}{Name of the term when no precision variable is used.}
}
\value{
Dataframe in a useable format for \code{rsample::bootstraps}.
}
\description{
Compute precision efficiency.
}
\details{
Compare the variance obtained when using a precision variable or not.
}
\examples{
# An example can be found in the location identified in the
# source section above at the github site
# https://github.com/FrankLef/FundamentalsCausalInference.
}
