% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/precision.R
\name{precision_stats}
\alias{precision_stats}
\alias{bootprecision.r}
\title{Compute Stats on Precision Efficiency}
\usage{
precision_stats(data, models, times = 1000, seed = NULL, none.name = "None")

bootprecision.r(data, models, times = 1000, seed = NULL, none.name = "None")
}
\arguments{
\item{data}{Dataframe of raw data.}

\item{models}{List of models with specs.}

\item{times}{Nb of boostrap iterations.}

\item{seed}{Boostrap seed.}

\item{none.name}{Name of the term when no precision variable is used.}
}
\value{
List with original dataframe of stats and original \code{models}.
}
\description{
Compute stats on precision efficiency.
}
\details{
Compute stats on precision efficiency using bootstrapping.
}
\examples{
# An example can be found in the location identified in the
# source section above at the github site
# https://github.com/FrankLef/FundamentalsCausalInference.
}
\seealso{
precision_eff
}
