% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prob_cond.R
\name{prob_cond}
\alias{prob_cond}
\alias{bootc}
\title{Compute estimates of the conditional association measures}
\usage{
prob_cond(
  data,
  formula = Y ~ T + A + H,
  cond0 = Y ~ A + H,
  cond1 = Y ~ T + A + H,
  family = c("binomial", "poisson", "gaussian"),
  R = 1000,
  conf = 0.95
)

bootc(
  data,
  formula = Y ~ T + A + H,
  cond0 = Y ~ A + H,
  cond1 = Y ~ T + A + H,
  family = c("binomial", "poisson", "gaussian"),
  R = 1000,
  conf = 0.95
)
}
\arguments{
\item{data}{Dataframe of raw data.}

\item{formula}{Formula of linear model.}

\item{cond0}{Formula of condition 0.}

\item{cond1}{Formula of condition 1.}

\item{family}{Family used by \code{glm}. Default is "binomial".}

\item{R}{Number of bootstrap replicates.}

\item{conf}{Confidence interval.}
}
\value{
Dataframe of effect measures.
}
\description{
Compute estimates of the conditional association measures.
}
\details{
Estimate the expected conditional outcomes and the
conditional effect or association measures.
IMPORTANT: This is the function in chapter 3 called lmodboot.r
It has been renamed bootc.r to avoid conflict with
lmodboot.r of chapter 2
}
