% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prob_lmod.R
\name{prob_lmod}
\alias{prob_lmod}
\alias{lmodboot}
\title{Estimate s sampling distribution by Bootstrapping}
\usage{
prob_lmod(data, formula = Y ~ T + A + H, condition.names = NULL)

lmodboot(data, formula = Y ~ T + A + H, condition.names = NULL)
}
\arguments{
\item{data}{Dataframe of raw data.}

\item{formula}{The model formula.}

\item{condition.names}{Character vector of conditioned variable names. If
\code{NULL} all independent variables will be used.}
}
\value{
Numeric vector of summarized results
}
\description{
Estimate s sampling distribution by bootstrapping.
}
