% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prob_lmod.R
\name{prob_lmodX}
\alias{prob_lmodX}
\title{Estimate s sampling distribution by Bootstrapping}
\usage{
prob_lmodX(
  data,
  formula = Y ~ T + A + H,
  cond = Y ~ T + A + H,
  R = 1000,
  conf = 0.95
)
}
\arguments{
\item{data}{Dataframe of raw data.}

\item{formula}{Formula of linear model.}

\item{cond}{Formula to identify conditioned variables.}

\item{R}{Number of bootstrap replicates.}

\item{conf}{Confidence interval.}
}
\value{
Vector of summarized results
}
\description{
Estimate s sampling distribution by bootstrapping.
}
