% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prop_mod.R
\name{prop_scores}
\alias{prop_scores}
\title{Fit the Propensity Score Model}
\source{
Section 6.2.2, p. 120-121.
}
\usage{
prop_scores(data, formula)
}
\arguments{
\item{data}{Data.frame.}

\item{formula}{Formula used to fit the propensity score.}
}
\value{
List with fitted values and the fit object.
}
\description{
Fit the propensity score model.
}
\details{
Fit the propensity score model to obtain the score for a given data set.
}
