% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data1.R
\docType{data}
\name{recovery}
\alias{recovery}
\title{RECOVERY trial of dexamethasone COVID-10 Collaborative Group}
\format{
Dataframe with 8 rows and 4 variables.
\describe{
\item{M}{1 = invasive mechanical ventilation prior to treatment}
\item{T}{1 = dexamethasone}
\item{Y}{1 = mortality}
\item{n}{number of observations}
}
}
\usage{
recovery
}
\description{
See section beginning of chapter 4.
}
\keyword{datasets}
