% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data1.R
\docType{data}
\name{sepsis}
\alias{sepsis}
\title{University of Florida Sepsis and Critical Illness (2017)}
\format{
Dataframe with 1190 rows and 4 variables.
\describe{
\item{gt65}{1 = greater than 65 years of age}
\item{shock}{1 = presence of septic shock at study entry}
\item{zubrod45}{1 = Zubrod score of 4 or 5 one year after study entry}
}
}
\usage{
sepsis
}
\description{
See exercise 1 of chapter 4.
}
\keyword{datasets}
