% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data1.R
\docType{data}
\name{sepsisb}
\alias{sepsisb}
\title{University of Florida Sepsis and Critical Illness (2017)}
\format{
Dataframe with 1190 rows and 4 variables.
\describe{
\item{shock}{1 = presence of septic shock at study entry}
\item{zubrodbase}{1 = Zubrod score of 3 or 4 at baseline}
\item{zubrod1yr}{1 = Zubrod score of 4 or 5 one year after study entry}
}
}
\usage{
sepsisb
}
\description{
See exercise 1 of chapter 7.
}
\keyword{datasets}
