% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_dag01.R
\name{sim_dag01}
\alias{sim_dag01}
\title{Simulate DAG # 1. Table 5.1.}
\usage{
sim_dag01(n = 1000, seed = 111)
}
\arguments{
\item{n}{Simulation size.}

\item{seed}{Integer,s seed.}
}
\value{
Dataframe
}
\description{
Simulate DAG # 1. Table 5.1.
}
\examples{
\dontrun{
sim_dag01()
}
}
