% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_doublewhatif.R
\name{sim_doublewhatif}
\alias{sim_doublewhatif}
\title{\code{doublewhatifsim} script rewritten}
\usage{
sim_doublewhatif(n = 1000, seed = 444)
}
\arguments{
\item{n}{Nb of observations.}

\item{seed}{Integer, the seed used for random numbers.}
}
\value{
Dataframe
}
\description{
\code{doublewhatifsim} script rewritten.
}
\details{
Simulate the What-If study data.
}
\examples{
\dontrun{
sim_doublewhatif()
}

}
