% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_intervals.R
\name{sim_intervals}
\alias{sim_intervals}
\alias{sim}
\title{Simulate a sampling distribution}
\usage{
sim_intervals(nsim = 10000, n = 10000, seed = 1009)

sim(nsim = 10000, n = 10000, seed = 1009)
}
\arguments{
\item{nsim}{Nb of replicates.}

\item{n}{Nb of observations.}

\item{seed}{Integer, the seed used for random numbers.}
}
\value{
List with percentage of bad and good.
}
\description{
Simulate a sampling distribution.
}
\details{
Simulate a sampling distribution using base R. Section 2.4.
}
\examples{
\dontrun{
sim_intervals()
}
\dontrun{
sim()
}
}
