% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mc_standdr.R
\name{standdr_est}
\alias{standdr_est}
\title{Estimates from Doubly Robust Standardization Simulation}
\usage{
standdr_est(Y, T, H)
}
\arguments{
\item{Y}{Vector of outcomes}

\item{T}{Vector of treatments}

\item{H}{Matrix of covariates}
}
\value{
List of estimates
}
\description{
Estimates from Doubly Robust Standardization Simulation
}
