% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mc_standdr.R
\name{standdr_sim}
\alias{standdr_sim}
\title{Data Simulation for Doubly Robust Standardization}
\usage{
standdr_sim(
  n = 3000,
  ss = 100,
  alpha = 0.13,
  beta = 20,
  probH = 0.05,
  seed = NULL
)
}
\arguments{
\item{n}{Number of individuals/observations.}

\item{ss}{Number of covariates.}

\item{alpha}{coefficient used to compute the distribution of \code{`T`}.}

\item{beta}{coefficient used to compute the distribution of \code{`T`}.}

\item{probH}{probability of H.}

\item{seed}{Seed value. default is \code{NULL}.}
}
\value{
List with a dataframe of Y, T and H and summary statitics.
}
\description{
Data Simulation for Doubly Robust Standardization
}
