% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mc_standdr.R
\name{standdr_stats}
\alias{standdr_stats}
\title{Compute Statistics from \code{standdr_sim}.}
\usage{
standdr_stats(x)
}
\arguments{
\item{x}{Vector of numeric values.}
}
\value{
list of statistics: \code{sun(x), mean(x), min(x), max(x)}.
}
\description{
Compute Statistics from \code{standdr_sim}.
}
