% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/time_msm.R, R/time_snmm.R
\name{time_msm}
\alias{time_msm}
\alias{msm.r}
\alias{snmm.r}
\title{Estimate Using Marginal Structural Models}
\source{
Section 13.1
}
\usage{
time_msm(data, outcome.name, exposure.names, confound.names)

msm.r(data, outcome.name, exposure.names, confound.names)

snmm.r(data, outcome.name, exposure.names, confound.names)
}
\arguments{
\item{data}{Dataframe.}

\item{outcome.name}{Name of outcome variable.}

\item{exposure.names}{Name of exposure variables. There must be 2 of them.}

\item{confound.names}{Names of the confound variables.}
}
\value{
Dataframe in a useable format for \code{rsample::bootstraps}.
}
\description{
Estimate using marginal structural models.
}
\details{
Estimate using marginal structural models with he tidyverse way of coding.
}
\examples{
# An example can be found in the location identified in the
# source section above at the github site
# https://github.com/FrankLef/FundamentalsCausalInference.
}
