% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/time_odtr.R
\name{time_odtr}
\alias{time_odtr}
\title{Optimal Dynamic Treatment Regime: All steps}
\source{
Section 13.3
}
\usage{
time_odtr(data, outcome.name, A1, A2, H2)
}
\arguments{
\item{data}{Dataframe.}

\item{outcome.name}{Name of outcome variable.}

\item{A1}{Name of A1 variable.}

\item{A2}{Name of A2 variable.}

\item{H2}{Name of H2 variable.}
}
\value{
Dataframe in a useable format for \code{rsample::bootstraps}.
}
\description{
Optimal dynamic treatment regime: All steps.
}
\details{
Optimal dynamic treatment regime: All steps, extract the optimal regime.
}
\examples{
# An example can be found in the location identified in the
# source section above at the github site
# https://github.com/FrankLef/FundamentalsCausalInference.
}
