% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/time_odtr.R
\name{time_odtr_optA2}
\alias{time_odtr_optA2}
\alias{A2opt.r}
\title{Optimal Dynamic Treatment Regime: Step 2}
\source{
Section 13.3
}
\usage{
time_odtr_optA2(data, A1, A2, H2)

A2opt.r(data, A1, A2, H2)
}
\arguments{
\item{data}{Dataframe.}

\item{A1}{Name of A1 variable.}

\item{A2}{Name of A2 variable.}

\item{H2}{Name of H2 variable.}
}
\value{
Dataframe.
}
\description{
Optimal dynamic treatment regime: Step 2.
}
\details{
Optimal dynamic treatment regime: Step 2, compute the optimal proportions of
A2.
}
\examples{
# An example can be found in the location identified in the
# source section above at the github site
# https://github.com/FrankLef/FundamentalsCausalInference.
}
