% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/time_odtr.R
\name{time_odtr_optimal}
\alias{time_odtr_optimal}
\alias{optimal.r}
\title{Optimal Dynamic Treatment Regime: Step 4}
\source{
Section 13.3
}
\usage{
time_odtr_optimal(data, A1, A2, H2)

optimal.r(data, A1, A2, H2)
}
\arguments{
\item{data}{Dataframe.}

\item{A1}{Name of A1 variable.}

\item{A2}{Name of A2 variable.}

\item{H2}{Name of H2 variable.}
}
\value{
Dataframe in a useable format for \code{rsample::bootstraps}.
}
\description{
Optimal dynamic treatment regime: Step 4.
}
\details{
Optimal dynamic treatment regime: Step 4, extract the optimal regime.
}
\examples{
# An example can be found in the location identified in the
# source section above at the github site
# https://github.com/FrankLef/FundamentalsCausalInference.
}
