% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/time_snmm.R
\name{time_snmm}
\alias{time_snmm}
\title{Estimate Using Structural Nested Mean Models}
\source{
Section 13.2
}
\usage{
time_snmm(data, outcome.name, exposure.names, confound.names)
}
\arguments{
\item{data}{Dataframe.}

\item{outcome.name}{Name of outcome variable.}

\item{exposure.names}{Name of exposure variables. There must be 2 of them.}

\item{confound.names}{Names of the confound variables.}
}
\value{
Dataframe in a useable format for \code{rsample::bootstraps}.
}
\description{
Estimate using structural nested mean models.
}
\details{
Estimate using structural nested mean models with he tidyverse way of coding.
}
\examples{
# An example can be found in the location identified in the
# source section above at the github site
# https://github.com/FrankLef/FundamentalsCausalInference.
}
