% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data1.R
\docType{data}
\name{whatif2dat}
\alias{whatif2dat}
\title{What-If study (Cook et al (2019)) with extended data}
\format{
Dataframe with 165 rows and 11 variables.
\describe{
\item{vl0}{Viral load at time 0}
\item{vlcont0}{Viral load at time 0}
\item{artad0}{todo}
\item{Vl4}{Viral load after 4 months}
\item{Vlcont4}{Viral load after 4 months}
\item{artad4}{todo}
\item{audit0}{todo}
\item{T}{1 = naltrexone}
\item{A}{1 = reduced drinking}
\item{lvlcont0}{todo}
\item{lvlcont4}{todo}
}
}
\usage{
whatif2dat
}
\description{
See section 1.2.3.1 for details.
}
\keyword{datasets}
