% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data1.R
\docType{data}
\name{whatifdat}
\alias{whatifdat}
\title{What-If study (Cook et al (2019))}
\format{
Dataframe with 165 rows and 4 variables.
\describe{
\item{T}{1 = naltrexone, 0 = placebo}
\item{A}{1 = reduced drinking for prior 30 days}
\item{H}{1 = unsuppressed HIV load}
\item{Y}{1 = unsuppressed HIV load at month 4}
}
}
\usage{
whatifdat
}
\description{
See section 1.2.3.1 for details.
}
\keyword{datasets}
