% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{audit_formula}
\alias{audit_formula}
\title{Audit the Formula Used for Causal Inference}
\usage{
audit_formula(data, formula, exposure.name, nvars = NA_integer_)
}
\arguments{
\item{data}{Dataframe of raw data.}

\item{formula}{Formula with at least 1 exposure and 1 extra variable.}

\item{exposure.name}{The name of the exposure variable.}

\item{nvars}{The required number of extra variables. If \code{NA_integer_}
then the nb of variables must be >= 1. The default value is \code{NA_integer_}.}
}
\value{
List with outcome.name, exposure.name and extra.names of variables.
}
\description{
Audit the formula used for causal inference.
}
\details{
The formula and the variables used must be validated to avoid many different
problems. All formula are assumed to have at least 1 outcome variable,
1 exposure variable and 1 extra variable. The extra variables can be confounds,
instrumental variable, etc. depending on the context.
}
