% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/backdr_dr.R
\name{backdr_dr}
\alias{backdr_dr}
\alias{standdr}
\title{Compute the doubly robust standardized estimates}
\usage{
backdr_dr(
  data,
  outcome.name = "Y",
  exposure.name = "T",
  confound.names = c("A", "H"),
  family = c("binomial", "poisson", "gaussian")
)

standdr(
  data,
  outcome.name = "Y",
  exposure.name = "T",
  confound.names = c("A", "H"),
  family = c("binomial", "poisson", "gaussian")
)
}
\arguments{
\item{data}{Dataframe of data.}

\item{outcome.name}{Name of outcome variable.}

\item{exposure.name}{Name of exposure variable.}

\item{confound.names}{Name of confound variable.}

\item{family}{family used for lm fit.}
}
\value{
Dataframe in a useable format for \code{rsample::bootstraps}.
}
\description{
Compute the doubly robust standardized estimates.
}
\details{
Compute the doubly robust standardized estimates using the code from section
6.3.
}
