% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/backdr_dr.R
\name{backdr_dr}
\alias{backdr_dr}
\alias{standdr}
\title{Compute the doubly robust standardized estimates}
\usage{
backdr_dr(
  data,
  formula = Y ~ T + H,
  family = c("binomial", "poisson", "gaussian"),
  R = 1000,
  conf = 0.95
)

standdr(
  data,
  formula = Y ~ T + H,
  family = c("binomial", "poisson", "gaussian"),
  R = 1000,
  conf = 0.95
)
}
\arguments{
\item{data}{Dataframe of raw data.}

\item{formula}{Formula in format \code{Y ~ T + ...} see details above.}

\item{family}{Character. Name of the distribution. Default is "binomial".}

\item{R}{Number of bootstrap replicates.}

\item{conf}{Confidence interval.}
}
\value{
Dataframe of estimates
}
\description{
Compute the doubly robust standardized estimates.
}
\details{
Compute the doubly robust standardized estimates using the code from section
6.3.
}
