% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/backdr_dr.R
\name{backdr_dr}
\alias{backdr_dr}
\alias{standdr}
\title{Compute the doubly robust standardized estimates}
\usage{
backdr_dr(
  data,
  formula = Y ~ T + A + H,
  exposure.name = "T",
  family = c("binomial", "poisson", "gaussian")
)

standdr(
  data,
  formula = Y ~ T + A + H,
  exposure.name = "T",
  family = c("binomial", "poisson", "gaussian")
)
}
\arguments{
\item{data}{Dataframe of raw data.}

\item{formula}{Formula representing the model.}

\item{exposure.name}{Name of exposure variable. All other independent
variables in the formula will be assumed to be confounds.}

\item{family}{Name of the model's family. Must be one of
\code{c("binomial", "poisson", "gaussian")}. default is \code{"binomial"}.}
}
\value{
Dataframe in a useable format for \code{rsample::bootstraps}.
}
\description{
Compute the doubly robust standardized estimates.
}
\details{
Compute the doubly robust standardized estimates using the code from section
6.3.
}
