% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/backdr_dr_bad.R
\name{backdr_dr_bad}
\alias{backdr_dr_bad}
\alias{badstanddr}
\title{Doubly robust standardized estimates with misspecified outcome model}
\usage{
backdr_dr_bad(data, formula = Y ~ T + H, exposure.name = "T")

badstanddr(data, formula = Y ~ T + H, exposure.name = "T")
}
\arguments{
\item{data}{Dataframe of raw data.}

\item{formula}{Formula representing the model.}

\item{exposure.name}{Name of exposure variable. All other independent
variables in the formula will be assumed to be confounds.}
}
\value{
Dataframe in a useable format for \code{rsample::bootstraps}.
}
\description{
Doubly robust standardized estimates with misspecified outcome model.
}
\details{
Compute the doubly robust standardized estimates using the code from section
6.3.
}
