% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/backdr_dr_bad.R
\name{backdr_dr_bad}
\alias{backdr_dr_bad}
\alias{badstanddr}
\title{Doubly robust standardized estimates with misspecified outcome model}
\usage{
backdr_dr_bad(
  data,
  outcome.name = "Y",
  exposure.name = "T",
  confound.names = "H"
)

badstanddr(data, outcome.name = "Y", exposure.name = "T", confound.names = "H")
}
\arguments{
\item{data}{Dataframe of data.}

\item{outcome.name}{Name of outcome variable.}

\item{exposure.name}{Name of exposure variable.}

\item{confound.names}{Name of confound variable.}
}
\value{
Dataframe in a useable format for \code{rsample::bootstraps}.
}
\description{
Doubly robust standardized estimates with misspecified outcome model.
}
\details{
Compute the doubly robust standardized estimates using the code from section
6.3.
}
