% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/backdr_dr_bad.R
\name{backdr_dr_bad}
\alias{backdr_dr_bad}
\alias{badstanddr}
\title{Doubly robust standardized estimates with misspecified outcome model}
\usage{
backdr_dr_bad(data, formula = Y ~ T + H, R = 1000, conf = 0.95)

badstanddr(data, formula = Y ~ T + H, R = 1000, conf = 0.95)
}
\arguments{
\item{data}{Dataframe of raw data.}

\item{formula}{Formula in format \code{Y ~ T + ...} see details above.}

\item{R}{Number of bootstrap replicates.}

\item{conf}{Confidence interval.}
}
\value{
Dataframe of estimates
}
\description{
Doubly robust standardized estimates with misspecified outcome model.
}
\details{
Compute the doubly robust standardized estimates using the code from section
6.3
}
