% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/backdr_exp.R
\name{backdr_exp}
\alias{backdr_exp}
\alias{standexp.r}
\title{Compute Standardized Estimates With Parametric Exposure Model}
\source{
Section 6.2.2
}
\usage{
backdr_exp(data, formula, exposure.name, confound.names, att = FALSE)

standexp.r(data, formula, exposure.name, confound.names, att = FALSE)
}
\arguments{
\item{data}{Dataframe of raw data.}

\item{formula}{Formula representing the model.}

\item{exposure.name}{Name of exposure variable.}

\item{confound.names}{Names of the confound variables.}

\item{att}{if \code{FALSE} calculate the standardized (unconfounded)
causal effect. If \code{TRUE} calculate the average effect of treatment
on the treated.}
}
\value{
Dataframe in a useable format for \code{rsample::bootstraps}.
}
\description{
Compute standardized estimates with parametric exposure model.
}
\details{
The standardized estimates are computed using the exposure model.
This method requires 2 different formulas which are created from the
arguments \code{formula}. The 2 formulas created are for the exposure model
and another one for the weighted linear model. Also \code{T}, i.e. the
exposure, must always be binary, if not it can be made binary
"one can first recode the data so that T = 1 when it is previously equaled,
and T = 0 when it previously equaled any value other than t", p. 113.
}
\examples{
# An example can be found in the location identified in the
# source section above at the github site
# https://github.com/FrankLef/FundamentalsCausalInference.
}
