% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/backdr_exp_gee.R
\name{backdr_exp_gee}
\alias{backdr_exp_gee}
\alias{standexpgee}
\title{Compute standardized estimates with parametric exposure model}
\usage{
backdr_exp_gee(
  data,
  outcome.name = "Y",
  exposure.name = "T",
  confound.names = "H"
)

standexpgee(
  data,
  outcome.name = "Y",
  exposure.name = "T",
  confound.names = "H"
)
}
\arguments{
\item{data}{Dataframe of data.}

\item{outcome.name}{Name of outcome variable.}

\item{exposure.name}{Name of exposure variable.}

\item{confound.names}{Name of confound variable.}
}
\value{
Dataframe of estimates
}
\description{
Compute standardized estimates with parametric exposure model.
}
\details{
The standardized estimates are computed using the exposure model and the
\code{geeglm} from the \code{gee} package.
This method requires 2 different formulas which are created from the
arguments \code{formula}. The 2 formulas created are for the exposure model
and another one for the weighted linear model.
}
\seealso{
backdr_exp
}
