% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/backdr_exp_np.R
\name{backdr_exp_np}
\alias{backdr_exp_np}
\alias{calc_exposure}
\title{Compute standardized averages using exposure modeling, non-parametric}
\usage{
backdr_exp_np(
  data,
  outcome.name = "Y",
  exposure.name = "T",
  confound.names = "H",
  att = FALSE
)

calc_exposure(
  data,
  outcome.name = "Y",
  exposure.name = "T",
  confound.names = "H",
  att = FALSE
)
}
\arguments{
\item{data}{Dataframe of data.}

\item{outcome.name}{Name of outcome variable.}

\item{exposure.name}{Name of exposure variable.}

\item{confound.names}{Name of confound variable.}

\item{att}{if \code{FALSE} calculate the standardized (unconfounded)
causal effect. If \code{TRUE} calculate the average effect of treatment
on the treated.}
}
\value{
Estimate using exposure-model standardization.
}
\description{
Compute standardized averages using exposure modeling, non-parametric.
}
\details{
Compute standardized averages using exposure modeling as described in
section 6.2.
}
