% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/backdr_exp_np.R
\name{backdr_exp_np}
\alias{backdr_exp_np}
\alias{mk.mortdat}
\alias{attsem.r}
\title{Compute Standardized Averages Using Exposure Modeling, Non Parametric}
\usage{
backdr_exp_np(data, formula = Y ~ T + H, exposure.name = "T", att = FALSE)

mk.mortdat(data, formula = Y ~ T + H, exposure.name = "T", att = FALSE)

attsem.r(data, formula = Y ~ T + H, exposure.name = "T", att = FALSE)
}
\arguments{
\item{data}{Dataframe of raw data.}

\item{formula}{Formula representing the model.}

\item{exposure.name}{Name of exposure variable. All other independent
variables in the formula will be assumed to be confounds.}

\item{att}{if \code{FALSE} calculate the standardized (unconfounded)
causal effect. If \code{TRUE} calculate the average effect of treatment
on the treated.}
}
\value{
Dataframe in a useable format for \code{rsample::bootstraps}.
}
\description{
Compute standardized averages using exposure modeling, non-parametric.
}
\details{
Compute standardized averages using exposure modeling. See introduciton of
section 6.2 and section 6.2.1.
}
