% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/backdr_exp_np.R
\name{backdr_exp_np}
\alias{backdr_exp_np}
\alias{calc_exposure}
\title{Compute standardized averages using exposure modeling, non-parametric}
\usage{
backdr_exp_np(
  data,
  outcome,
  exposure,
  confound,
  att = FALSE,
  R = 1000,
  conf = 0.95
)

calc_exposure(
  data,
  outcome,
  exposure,
  confound,
  att = FALSE,
  R = 1000,
  conf = 0.95
)
}
\arguments{
\item{data}{Dataframe}

\item{outcome}{Outcome variable}

\item{exposure}{Exposure variable}

\item{confound}{Confound variable}

\item{att}{if \code{FALSE} calculate the standardized (unconfounded)
causal effect. If \code{TRUE} calculate the average effect of treatment
on the treated.}

\item{R}{Number of bootstrap replicates.}

\item{conf}{Confidence interval.}
}
\value{
Estimate using exposure-model standardization.
}
\description{
Compute standardized averages using exposure modeling, non-parametric.
}
\details{
Compute standardized averages using exposure modeling as described in
section 6.2.
}
