% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/backdr_exp_np.R
\name{backdr_exp_np}
\alias{backdr_exp_np}
\alias{calc_exposure}
\title{Compute standardized averages using exposure modeling, non-parametric}
\usage{
backdr_exp_np(data, formula = Y ~ T + H, weights = "n")

calc_exposure(data, formula = Y ~ T + H, weights = "n")
}
\arguments{
\item{data}{Dataframe}

\item{formula}{Formula, must be in the format \code{Y ~ `T` + H}, i.e.
only 1 covariate H.}

\item{weights}{String. Name of the columns with the weights that will
be used to create probabilities summing up to 1.}
}
\value{
List with 3 elements: EY1, EY0, EY0T1. See section 6.2.
for more details.
}
\description{
Compute standardized averages using exposure modeling, non-parametric.
}
\details{
Compute standardized averages using exposure modeling as described in
section 6.2.
}
