% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/backdr_out_np.R
\name{backdr_out_np}
\alias{backdr_out_np}
\alias{stand}
\title{Compute standardized estimates via Outcome Modeling, Non-Parametric}
\usage{
backdr_out_np(
  data,
  formula = Y ~ T + H + T * H,
  att = FALSE,
  R = 1000,
  conf = 0.95
)

stand(data, formula = Y ~ T + H + T * H, att = FALSE, R = 1000, conf = 0.95)
}
\arguments{
\item{data}{Dataframe of raw data.}

\item{formula}{Formula in format \code{Y ~ T + ...}, see details.}

\item{att}{if \code{FALSE} calculate the standardized (unconfounded)
causal effect. If \code{TRUE} calculate the average effect of treatment
on the treated.}

\item{R}{Number of bootstrap replicates.}

\item{conf}{Confidence interval.}
}
\value{
Estimate using outcome-model standardization.
}
\description{
Compute standardized estimates via outcome modeling, non-parametric.
}
\details{
The standardized estimates are computed using the non-parametric outcome
model. IMPORTANT: The formula must be in the format \code{Y ~ T + ...} where the
exposure is in the first position \code{T}.
}
