% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/backdr_out_np.R
\name{backdr_out_np}
\alias{backdr_out_np}
\title{Standardized estimates via Outcome Modeling, Non-Parametric}
\source{
Section 6.1
}
\usage{
backdr_out_np(data, formula, exposure.name, confound.names, att = FALSE)
}
\arguments{
\item{data}{Dataframe of raw data.}

\item{formula}{Formula representing the model.}

\item{exposure.name}{Name of exposure variable.}

\item{confound.names}{Names of the confound variables.}

\item{att}{if \code{FALSE} calculate the standardized (unconfounded)
causal effect. If \code{TRUE} calculate the average effect of treatment
on the treated.}
}
\value{
Dataframe in a useable format for \code{rsample::bootstraps}.
}
\description{
Standardized estimates via outcome modeling, non-parametric.
}
\details{
The standardized estimates are computed using the non-parametric outcome
model. The calculations are performed without regression.
}
\examples{
# An example can be found in the location identified in the
# source section above at the github site
# https://github.com/FrankLef/FundamentalsCausalInference.
}
