% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/backdr_out_sat.R
\name{backdr_out_sat}
\alias{backdr_out_sat}
\alias{stand.r}
\title{Standardized estimates via Outcome Modeling, Saturated regression}
\usage{
backdr_out_sat(data, formula = Y ~ A * H, exposure.name = "A", att = FALSE)

stand.r(data, formula = Y ~ A * H, exposure.name = "A", att = FALSE)
}
\arguments{
\item{data}{Dataframe of raw data.}

\item{formula}{Formula representing the model.}

\item{exposure.name}{Name of exposure variable. All other independent
variables in the formula will be assumed to be confounds.}

\item{att}{if \code{FALSE} calculate the standardized (unconfounded)
causal effect. If \code{TRUE} calculate the average effect of treatment
on the treated.}
}
\value{
Dataframe in a useable format for \code{rsample::bootstraps}.
}
\description{
Standardized estimates via outcome modeling, saturated regression.
}
\details{
The standardized estimates are computed using a saturated regression fit.
That is all variables and their interactions are used. \emph{make sure all
variables and their interactions are included in the formula}.
See p. 101.
}
