% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boot_utils.R
\name{boot_est}
\alias{boot_est}
\title{Get estimate and CI using a function as input}
\usage{
boot_est(
  data,
  func,
  times = 1000,
  alpha = 0.05,
  inv = c("exp", "expit", "none"),
  evars = c("standard", "modifier", "logit"),
  ...
)
}
\arguments{
\item{data}{Dataframe of raw data.}

\item{func}{Function applied to data by bootstrapping.}

\item{times}{Number of bootstrap replicates. Default is 1000.}

\item{alpha}{Alpha used by percentile to give interval in
\code{c(alpha, 1- alpha)}.}

\item{inv}{Choice of inverse function to apply to the effect measure
\code{exp} will exponentiate the result (Default), \code{expit} will apply
the inverse logit and \code{none} will do nothing (identity function).}

\item{evars}{String. Should one of \code{c("standard", "modifier", "logit")}.
Default value is \code{"standard"}.}

\item{...}{Other named arguments for \code{func}.}
}
\value{
Dataframe of estimates with confidence interval..
}
\description{
Get estimate and CI using a function as input.
}
\details{
Boostrap with a function and extra arguments as input. The function is called
\code{statistic} to remain constant with \code{boot::boot}.
}
