% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boot_utils.R
\name{boot_run}
\alias{boot_run}
\title{Bootstrap and generate a dataframe of estimates with CI}
\usage{
boot_run(data, statistic, R = 1000, conf = 0.95, ...)
}
\arguments{
\item{data}{Dataframe of raw data.}

\item{statistic}{Function applied to data by bootstrapping.}

\item{R}{Number of bootstrap replicates. Default is 1000.}

\item{conf}{Confidence interval width. Default is 0.95.}

\item{...}{Other named arguments for \code{statistic}.}
}
\value{
Dataframe of estimates with CI.
}
\description{
Bootstrap and generate a dataframe of estimates with CI.
}
\details{
Generate a dataframe of estimates with the columns
\code{c("est", "conf", "lci", "uci")}. The \code{boot::boot} function is used
for bootstraping and the function \code{boot::boot.ci} is used to compute the
confidence interval.
}
\seealso{
boot::boot boot::boot.ci
}
