% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/did_linear.R
\name{did_linear}
\alias{did_linear}
\alias{didlinear}
\title{Compute the DiD Estimator with a Linear Model}
\source{
Section 7.1.1
}
\usage{
did_linear(
  data,
  formula,
  exposure.name,
  confound.names,
  names_to = "var",
  timevar = "time"
)

didlinear(
  data,
  formula,
  exposure.name,
  confound.names,
  names_to = "var",
  timevar = "time"
)
}
\arguments{
\item{data}{Dataframe of raw data.}

\item{formula}{Formula representing the model.}

\item{exposure.name}{Name of exposure variable.}

\item{confound.names}{Names of the confound variables.}

\item{names_to}{Name that will be used for the name variable.}

\item{timevar}{Name that will be used for the time variable.}
}
\value{
Dataframe in a useable format for \code{rsample::bootstraps}.
}
\description{
Compute the DiD Estimator with a linear model.
}
\details{
The DiD Estimator using the scripts from section 7.2, p. 141-142.
}
\seealso{
did_longer
}
