% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/did_linear.R
\name{did_linear}
\alias{did_linear}
\title{Compute the DiD Estimator with a Linear Model}
\usage{
did_linear(
  data,
  outcome.name = "Y1",
  exposure.name = "T",
  confound.names = "Y0",
  names_to = "var",
  timevar = "time"
)
}
\arguments{
\item{data}{Dataframe of raw data.}

\item{outcome.name}{Name of final outcome.}

\item{exposure.name}{Name of exposure variable.}

\item{confound.names}{Name of baseline outcome (confound).}

\item{names_to}{Name that will be used for the name variable.}

\item{timevar}{Name that will be used for the time variable.}
}
\value{
Estimates using the difference in differences.
}
\description{
Compute the DiD Estimator with a linear model.
}
\details{
The DiD Estimator using the scripts from section 7.2, p. 141-142.
}
\seealso{
did_longer
}
