% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/did_logistic.R
\name{did_logistic}
\alias{did_logistic}
\alias{didlogistic}
\title{Compute the DiD Estimator with a Lofistic Model}
\usage{
did_logistic(
  data,
  formula,
  exposure.name,
  confound.names,
  names_to = "var",
  timevar = "time"
)

didlogistic(
  data,
  formula,
  exposure.name,
  confound.names,
  names_to = "var",
  timevar = "time"
)
}
\arguments{
\item{data}{Dataframe of raw data.}

\item{formula}{Formula representing the model.}

\item{exposure.name}{Name of exposure variable.}

\item{confound.names}{Names of the confound variables.}

\item{names_to}{Name that will be used for the name variable.}

\item{timevar}{Name that will be used for the time variable.}
}
\value{
Dataframe in a useable format for \code{rsample::bootstraps}.
}
\description{
Compute the DiD Estimator with a logistic model.
}
\details{
the DiD Estimator using the scripts from section 7.2, p. 141-142.
}
\seealso{
did_longer
}
