% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/did_loglinear.R
\name{did_loglinear}
\alias{did_loglinear}
\title{Compute the DiD Estimator with a Loglinear Model}
\usage{
did_loglinear(
  data,
  outcomes = c("Y0", "Y1"),
  outcome = "Y",
  treatment = "T",
  names_to = "var",
  timevar = "time",
  R = 1000,
  conf = 0.95
)
}
\arguments{
\item{data}{Dataframe of raw data.}

\item{outcomes}{Character vector of base and final variables.}

\item{outcome}{Name that will be used for the single outcome variable.}

\item{treatment}{Name of the treatment variable.}

\item{names_to}{Name that will be used for the name variable.}

\item{timevar}{Name that will be used for the time variable.}

\item{R}{Number of bootstrap replicates. Default is 1000.}

\item{conf}{Confidence interval. Default is 0.95.}
}
\value{
Estimates using the difference in differences.
}
\description{
Compute the DiD Estimator with a loglinear model.
}
\details{
the DiD Estimator using the scripts from section 7.2, p. 141-142.
}
\seealso{
did_longer
}
