% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{fci_tbl_03_02}
\alias{fci_tbl_03_02}
\title{Table 3.2 data}
\format{
Dataframe with 6 rows and 5 variables.
\describe{
\item{estimator}{name of estimator}
\item{est}{estimate}
\item{conf}{confidence interval width}
\item{lci}{lower confidence interval}
\item{uci}{upper confidence interval}
}
}
\usage{
fci_tbl_03_02
}
\description{
See chapter 3.
}
\keyword{datasets}
