% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data2.R
\docType{data}
\name{fci_tbl_04_02}
\alias{fci_tbl_04_02}
\title{Table 4.2}
\format{
Dataframe with 18 rows and 6 variables.
\describe{
\item{estimator}{name of estimator}
\item{group}{grouping variable}
\item{est}{estimate}
\item{conf}{confidence interval width}
\item{lci}{lower confidence interval}
\item{uci}{upper confidence interval}
}
}
\usage{
fci_tbl_04_02
}
\description{
See chapter 4.
}
\keyword{datasets}
