% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data2.R
\docType{data}
\name{fci_tbl_07_02}
\alias{fci_tbl_07_02}
\title{Table 7.2}
\format{
Dataframe with 13 rows and 8 variables.
\describe{
\item{model}{model used for estimation}
\item{term}{name of estimator}
\item{Truth}{true value}
\item{.estimate}{estimate}
\item{.lower}{lower confidence interval}
\item{.upper}{upper confidence interval}
\item{.alpha}{alpha value confidence interval width = 1 - alpha}
\item{.method}{method used used to estimate interval}
}
}
\usage{
fci_tbl_07_02
}
\description{
See chapter 7, section 7.2.
}
\keyword{datasets}
