% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data2.R
\docType{data}
\name{fci_tbl_09_01a}
\alias{fci_tbl_09_01a}
\title{Table 9.1 using qt() instead of 1.96 for CI}
\format{
Dataframe with 10 rows and 8 variables.
\describe{
\item{model}{model used for estimation}
\item{term}{name of estimator}
\item{Truth}{true value}
\item{.estimate}{estimate}
\item{.lower}{lower confidence interval}
\item{.upper}{upper confidence interval}
\item{.alpha}{alpha value confidence interval width = 1 - alpha}
\item{.method}{method used used to estimate interval}
}
}
\usage{
fci_tbl_09_01a
}
\description{
See chapter 9, section 9.3.
}
\keyword{datasets}
