% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fciR-package.R
\docType{package}
\name{fciR}
\alias{fciR}
\alias{fciR-package}
\title{A package to study with Fundamentals of Causal Inference by Babette Brumback}
\source{
Fundamentals of Causal Inference with R, Babette A. Brumback,
2022, CRC Press
}
\description{
Pretty much all the functions developed by Babette Brumback in the book
Fundamentals of Causal Inference.
}
\section{fciR functions}{

The functions ...
}

\author{
\strong{Maintainer}: François Lefebvre \email{flefebvre01@hotmail.com} (\href{https://orcid.org/YOUR-ORCID-ID}{ORCID})

}
\keyword{internal}
