% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frontdr.R
\name{frontdr_np}
\alias{frontdr_np}
\title{Estimate the effect using Front-Door Method}
\usage{
frontdr_np(data, outcome.name, exposure.name, confound.names)
}
\arguments{
\item{data}{Dataframe of raw data.}

\item{outcome.name}{Name of outcome variable.}

\item{exposure.name}{Name of exposure variable.}

\item{confound.names}{Name of surrogate variable.}
}
\value{
Numeric vector.
}
\description{
Estimate the effect using Front-door method.
}
\details{
Estimate the effect using Front-door method. See chap. 8, section 8.3
for details.
}
