% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frontdr.R
\name{frontdr_np}
\alias{frontdr_np}
\title{Estimate and CI using the Front-Door Method}
\usage{
frontdr_np(data, outcome, exposure, surrogate, R = 1000, conf = 0.95)
}
\arguments{
\item{data}{Dataframe of raw data.}

\item{outcome}{Name of outcome variable.}

\item{exposure}{Name of exposure variable.}

\item{surrogate}{Name of surrogate variable.}

\item{R}{Number of bootstrap replicates.}

\item{conf}{Confidence interval.}
}
\value{
Dataframe of estimate and CI using the front-door method.
}
\description{
Estimate and CI using the front-door method.
}
\details{
Estimate and CI using the front-door method as described in section 8.3
of chater 8.
}
