% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggp_measures.R
\name{ggp_measures}
\alias{ggp_measures}
\title{Plot of effect measures}
\usage{
ggp_measures(data, title = "Title", subtitle = "Subtitle")
}
\arguments{
\item{data}{Data frame with effect measures.}

\item{title}{Title.}

\item{subtitle}{Subtitle.}
}
\value{
A ggplot object.•
}
\description{
Plot of effect measures.
}
\details{
Plot of effect measures with causal power (CF) and attributable fraction
(AF). See secion 3.3, p. 44-45 for more details.
}
