% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggp_measures.R
\name{ggp_measures}
\alias{ggp_measures}
\title{Plot of effect measures}
\usage{
ggp_measures(
  data,
  title = "Title",
  subtitle = "Subtitle",
  vline = list(colors = c("lightseagreen", "violet"), linetype = "solid", size = 4,
    alpha = 0.5),
  pointrange = list(size = 1, fatten = 3),
  text = list(size = 4, color = "navy", digits = 2),
  scale_color = list(zero = "darkgreen", one = "magenta")
)
}
\arguments{
\item{data}{Data frame with effect measures.}

\item{title}{Title.}

\item{subtitle}{Subtitle.}

\item{vline}{geom_vline specs.}

\item{pointrange}{geom_pointrange specs.}

\item{text}{geom_text_repel specs.}

\item{scale_color}{scale_color_manual color specs.}
}
\value{
A ggplot object.
}
\description{
Plot of effect measures.
}
\details{
Plot of effect measures with causal power (CF) and attributable fraction
(AF). See secion 3.3, p. 44-45 for more details.
}
