% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data1.R
\docType{data}
\name{gss}
\alias{gss}
\title{Dataset from 2018 GSS}
\format{
Dataframe with 2348 rows and 12 variables.
\describe{
\item{age}{todo}
\item{gt65}{todo}
\item{attend}{todo}
\item{gthsedu}{todo}
\item{magthsedu}{todo}
\item{pagthsedu}{todo}
\item{fair}{todo}
\item{owngun}{todo}
\item{conservative}{todo}
\item{trump}{todo}
\item{white}{todo}
\item{female}{todo}
}
an exported object
}
\source{
\url{www.gss.norc.org}
}
\usage{
gss
}
\description{
See section 1.2.4. Data on 2348 surveyed individuals.
}
\keyword{datasets}
