% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/instr_linear.R
\name{instr_linear}
\alias{instr_linear}
\alias{ividentity.r}
\title{Estimate Effect Using Instrument Variables}
\usage{
instr_linear(
  data,
  formula = Y ~ A * T,
  exposure.name = "A",
  tol = .Machine$double.eps^0.5
)

ividentity.r(
  data,
  formula = Y ~ A * T,
  exposure.name = "A",
  tol = .Machine$double.eps^0.5
)
}
\arguments{
\item{data}{Dataframe of raw data.}

\item{formula}{Formula representing the model.}

\item{exposure.name}{Name of exposure variable. The other independent
variables in the formula will be assumed to be the instrument variable.
There can be only one exposure variable and one instrument variable.}

\item{tol}{Numeric > 0. Tolerance used in estimation. Default is
.Machine$double.eps^0.5.}
}
\value{
Dataframe in a useable format for \code{rsample::bootstraps}.
}
\description{
Estimate effect using instrument variables.
}
\details{
See section 9.3 p. 167 for details.
}
