% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/instr_linear.R
\name{instr_linear}
\alias{instr_linear}
\title{Estimate Effect Using Instrument Variables}
\usage{
instr_linear(
  data,
  outcome.name = "Y",
  exposure.name = "A",
  instrument.name = "T",
  tol = .Machine$double.eps^0.5
)
}
\arguments{
\item{data}{Dataframe of raw data.}

\item{outcome.name}{Name of outcome variable.}

\item{exposure.name}{Name of exposure variable.}

\item{instrument.name}{Name of instrument variable.}

\item{tol}{Numeric > 0. Tolerance for denominator of IV. When the tolerance
fails the IV s replaced by \code{NA}}
}
\value{
Numeric vector of estimates.
}
\description{
Estimate effect using instrument variables.
}
\details{
See section 9.3 p. 167 for details.
}
