% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/instr_loglinear.R
\name{instr_loglinear}
\alias{instr_loglinear}
\alias{ivlog.r}
\title{Estimate Effect Using Instrument Variables via Logarithmic fit}
\usage{
instr_loglinear(
  data,
  formula = Y ~ A * T,
  exposure.name = "A",
  tol = .Machine$double.eps^0.5,
  niter = 10L
)

ivlog.r(
  data,
  formula = Y ~ A * T,
  exposure.name = "A",
  tol = .Machine$double.eps^0.5,
  niter = 10L
)
}
\arguments{
\item{data}{Dataframe of raw data.}

\item{formula}{Formula representing the model.}

\item{exposure.name}{Name of exposure variable. The other independent
variables in the formula will be assumed to be the instrument variable.
There can be only one exposure variable and one instrument variable.}

\item{tol}{Numeric > 0. Tolerance used in estimation. Default is
.Machine$double.eps^0.5.}

\item{niter}{Number of iterations}
}
\value{
Dataframe in a useable format for \code{rsample::bootstraps}.
}
\description{
Estimate effect using instrument variables via logarithmic fit.
}
\details{
See section 9.3 p. 167-168 for details.
}
