% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/instr_vars.R
\name{instr_vars}
\alias{instr_vars}
\alias{iv.r}
\title{Compute ITT, CACE and ATT from Instrument Variables}
\source{
Section 9.3
}
\usage{
instr_vars(
  data,
  formula,
  exposure.name,
  instrument.name,
  tol = .Machine$double.eps^0.5
)

iv.r(
  data,
  formula,
  exposure.name,
  instrument.name,
  tol = .Machine$double.eps^0.5
)
}
\arguments{
\item{data}{Dataframe of raw data.}

\item{formula}{Formula representing the model.}

\item{exposure.name}{Name of exposure variable.}

\item{instrument.name}{Name of instrument variable.}

\item{tol}{Numeric > 0. Tolerance used in estimation. Default is
.Machine$double.eps^0.5.}
}
\value{
Dataframe in a useable format for \code{rsample::bootstraps}.
}
\description{
Compute ITT, CACE and ATT from instrument variables.
}
\details{
See section 9.3 on p. 164 for details. When the tolerance is not met, the
}
