% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jack_utils.R
\name{jack_run}
\alias{jack_run}
\title{Estimate of Effect Measure and CI With Jacknife (LOO)}
\usage{
jack_run(data, func, conf = 0.95, ...)
}
\arguments{
\item{data}{Dataframe of raw data.}

\item{func}{Function to estimate the effect measure.}

\item{conf}{Confidence interval width. Default is 0.95.}

\item{...}{Other named arguments for \code{func}.}
}
\value{
Dataframe of estimates with confidence interval.
}
\description{
Estimate of effect measure and CI With jacknife (LOO).
}
\details{
Compute the confidence interval of effect measures using the jacknife method.
It uses the \code{loo_cv} function from the \code{rsample} package. The
\code{jacknife} funciton from the \code{resample} is not maintained anymore
and the author of \code{resample} recommend using the \code{rsample} package.
}
