% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meas_effect_cond.R
\name{meas_effect_cond}
\alias{meas_effect_cond}
\alias{bootc}
\title{Compute estimates of the conditional association measures}
\usage{
meas_effect_cond(
  data,
  outcome.name = "Y",
  exposure.name = "T",
  confound.names = c("A", "H"),
  family = c("binomial", "poisson", "gaussian")
)

bootc(
  data,
  outcome.name = "Y",
  exposure.name = "T",
  confound.names = c("A", "H"),
  family = c("binomial", "poisson", "gaussian")
)
}
\arguments{
\item{data}{Dataframe of raw data.}

\item{outcome.name}{Name of outcome variable.}

\item{exposure.name}{Name of exposure variable.}

\item{confound.names}{Character vector of confound variable names.}

\item{family}{Name of distribution family. Must be in
\code{c("binomial", "poisson", "gaussian")}. Default is \code{binomial}.}
}
\value{
Numeric vector of summarized results.
}
\description{
Compute estimates of the conditional association measures.
}
\details{
Estimate the expected conditional outcomes and the
conditional effect or association measures.
IMPORTANT: This is the function in chapter 3 called lmodboot.r
It has been renamed bootc.r to avoid conflict with
lmodboot.r of chapter 2
}
