% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meas_effect_modif.R
\name{meas_effect_modif}
\alias{meas_effect_modif}
\alias{boot}
\title{Compute estimates of the association measures for 2 strata}
\usage{
meas_effect_modif(
  data,
  outcome.name = "Y",
  exposure.name = "T",
  modifier.name = "M"
)

boot(data, outcome.name = "Y", exposure.name = "T", modifier.name = "M")
}
\arguments{
\item{data}{Dataframe of raw data.}

\item{outcome.name}{Name of outcome variable.}

\item{exposure.name}{Name of exposure variable.}

\item{modifier.name}{Name of modifier variable.}
}
\value{
Numeric vector of summarized results.
}
\description{
Compute estimates of the association measures for 2 strata.
}
\details{
The formula must always be in form \code{Y ~ `T` + M}, that is with only
2 predictors: \code{T} as the treatment variable and \code{M} as the modifier
variable.
}
