% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meas_effect_modif.R
\name{meas_effect_modif}
\alias{meas_effect_modif}
\alias{boot}
\title{Compute estimates of the association measures for 2 strata}
\usage{
meas_effect_modif(data, formula = Y ~ T + M, exposure.name = "T")

boot(data, formula = Y ~ T + M, exposure.name = "T")
}
\arguments{
\item{data}{Dataframe of raw data.}

\item{formula}{The model formula.}

\item{exposure.name}{Name of the exposure variable.}
}
\value{
Dataframe in a useable format for \code{rsample::bootstraps}.
}
\description{
Compute estimates of the association measures for 2 strata.
}
\details{
The formula must always be in form \code{Y ~ `T` + M}, that is with only
2 predictors: \code{T} as the treatment variable and \code{M} as the modifier
variable. The exposure variable is identified by the argument
\code{exposure.name}, the other variable is assumed to be the modifier
variable.
}
