% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meas_effect_uncond.R
\name{meas_effect_uncond}
\alias{meas_effect_uncond}
\alias{bootu}
\title{Compute estimates of the unconditional association measures}
\usage{
meas_effect_uncond(data, formula = Y ~ T, R = 1000, conf = 0.95)

bootu(data, formula = Y ~ T, R = 1000, conf = 0.95)
}
\arguments{
\item{data}{Dataframe of raw data.}

\item{formula}{Formula of linear model.}

\item{R}{Number of bootstrap replicates.}

\item{conf}{Confidence interval}
}
\value{
Dataframe of summarized results
}
\description{
Compute estimates of the unconditional association measures.
}
\details{
Compute estimates of the unconditional association measures,
hence the u in \code{bootu()}, and their confidence intervals.
1- We use Gaussian, Poisson and Binomial glm to solve the
estimating equations of the measures, not their distributions
2- The association measures' distributions (ci) are estimated by
bootstrapping
Assumptions: We assume that (3.2) holds but not (3.1)
See p. 45 and 46 for more details.
}
