% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prob_lmod.R
\name{prob_lmod}
\alias{prob_lmod}
\alias{lmodboot}
\title{Estimate s sampling distribution by Bootstrapping}
\usage{
prob_lmod(
  data,
  outcome.name = "Y",
  input.names = c("T", "A", "H"),
  condition.names = input.names
)

lmodboot(
  data,
  outcome.name = "Y",
  input.names = c("T", "A", "H"),
  condition.names = input.names
)
}
\arguments{
\item{data}{Dataframe of raw data.}

\item{outcome.name}{Name of outcome variable.}

\item{input.names}{Character vector of input variable names.}

\item{condition.names}{Character vector of conditioned variable names.}
}
\value{
Vector of summarized results
}
\description{
Estimate s sampling distribution by bootstrapping.
}
