% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prob_lmod.R
\name{prob_lmod_td}
\alias{prob_lmod_td}
\title{Estimate s sampling distribution by Bootstrapping}
\usage{
prob_lmod_td(data, formula = Y ~ T + A + H, condition.names = NULL)
}
\arguments{
\item{data}{Dataframe of raw data.}

\item{formula}{The model formula.}

\item{condition.names}{Character vector of conditioned variable names. If
\code{NULL} all independent variables will be used.}
}
\value{
Dataframe in a useable format for \code{rsample::bootstraps}.
}
\description{
Estimate s sampling distribution by bootstrapping.
}
\details{
Estimate a sampling distribution. See details in chapter 2, section 2.4.
The package \code{rsample} is used.
}
