% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prop_scores.R
\name{prop_scores}
\alias{prop_scores}
\alias{prop.r}
\title{Fit the Propensity Score Model}
\source{
Section 6.2.2, p. 120-121. Used throughout chapter 10.
}
\usage{
prop_scores(data, formula)

prop.r(data, formula)
}
\arguments{
\item{data}{Data.frame.}

\item{formula}{Formula used to fit the propensity score.}
}
\value{
List with
\describe{
\item{scores}{Vector of propensity scores.}
\item{fit}{The object of class glm returned by glm().}
}
}
\description{
Fit the propensity score model.
}
\details{
Fit the propensity score model to obtain the score for a given data set.
}
